Move Veryl comb construction into frontend
