+++ crates/celox/src/logic_tree/comb.rs
@@ -1,3281 +1,1 @@
-mod effect;
-mod expr;
-mod recover_unrolled;
-mod state;
-
-pub(crate) mod node {
-    pub use celox_slt::{
-        NodeId, SLTForEffect, SLTForFoldGroupState, SLTForUpdate, SLTIndex, SLTIndexKind,
-        SLTLoopBound, SLTNode, SLTNodeArena, SLTNodeArenaEditError, SLTStepOp,
-    };
-}
-
-pub(crate) mod node_facts {
-    pub use celox_slt::{SLTNodeFacts, SLTNodeFactsError};
-}
-
-pub use celox_slt::{LogicPath, LogicPathTarget};
-pub use state::{BoundaryMap, SymbolicStore};
-
-use std::{collections::BTreeSet, hash::Hash};
-
-use crate::ParserError;
-use crate::logic_tree::range_store::{RangeStore, RangeStoreError};
-use crate::parser::loop_provenance::LoopRecoveryCandidate;
-use crate::parser::{LoweringPhase, case::case_arm_condition_expr, resolve_total_width};
-use crate::{
-    HashMap, HashSet,
-    ir::{
-        BinaryOp, BitAccess, CombObserver, RuntimeEventKind, RuntimeEventSite, UnaryOp, VarAtomBase,
-    },
-    parser::bitaccess::{PartSelectGeometry, eval_constexpr, eval_var_select, select_geometry},
-};
-use num_bigint::{BigInt, BigUint, Sign};
-use num_traits::ToPrimitive as _;
-use veryl_analyzer::ir::{
-    ArrayLiteralItem, AssignStatement, CaseStatement, CombDeclaration, Expression, Factor,
-    ForBound, ForRange, ForStatement, IfStatement, Module, Op, Statement, SystemFunctionCall,
-    SystemFunctionInput, SystemFunctionKind, VarId, VarIndex, VarSelect,
-};
-use veryl_analyzer::value::{Value, byte_value_to_string};
-use veryl_parser::resource_table;
-use veryl_parser::token_range::TokenRange;
-
-use effect::{
-    CombEffectCollector, collect_comb_effects_statements, statements_contain_runtime_effect,
-    subtract_written_sensitivity,
-};
-pub(crate) use expr::coerce_node_width;
-use expr::{eval_array_literal_expression, eval_function_body_return, merge_boundaries};
-pub use expr::{eval_assignment_expression, eval_expression, get_width};
-use state::{FunctionControlState, LoopControlState};
-
-type ActiveGuard = (NodeId, HashSet<VarAtomBase<VarId>>);
-
-pub(crate) use node::SLTNodeArenaEditError;
-pub use node::{
-    NodeId, SLTForEffect, SLTForFoldGroupState, SLTForUpdate, SLTIndex, SLTIndexKind, SLTLoopBound,
-    SLTNode, SLTNodeArena, SLTStepOp,
-};
-pub use node_facts::{SLTNodeFacts, SLTNodeFactsError};
-
-pub(super) fn range_store_error(
-    context: &'static str,
-    error: RangeStoreError,
-    token: Option<&TokenRange>,
-) -> ParserError {
-    ParserError::illegal_context(context, error.to_string(), token)
-}
-
-#[cfg(test)]
-fn parse_comb(
-    module: &Module,
-    decl: &CombDeclaration,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<
-    (
-        Vec<LogicPath<VarId>>,
-        SymbolicStore<VarId>,
-        BoundaryMap<VarId>,
-        Vec<CombObserver<VarId>>,
-        Vec<RuntimeEventSite>,
-    ),
-    ParserError,
-> {
-    parse_comb_with_loop_recovery(module, decl, arena, &[])
-}
-
-pub(crate) fn parse_comb_with_loop_recovery(
-    module: &Module,
-    decl: &CombDeclaration,
-    arena: &mut SLTNodeArena<VarId>,
-    loop_candidates: &[LoopRecoveryCandidate],
-) -> Result<
-    (
-        Vec<LogicPath<VarId>>,
-        SymbolicStore<VarId>,
-        BoundaryMap<VarId>,
-        Vec<CombObserver<VarId>>,
-        Vec<RuntimeEventSite>,
-    ),
-    ParserError,
-> {
-    // 1. Initialization: Create a RangeStore for each variable in the module.
-    // Variables start in an 'unassigned' state (None), representing their initial input values.
-    let mut current_store = SymbolicStore::default();
-    for (id, var) in &module.variables {
-        let width = resolve_total_width(module, var)?;
-        current_store.insert(*id, RangeStore::new(None, width));
-    }
-
-    let mut written_accesses = HashMap::default();
-    collect_written_accesses(module, &decl.statements, &mut written_accesses)?;
-    let written_atoms: Vec<_> = written_accesses
-        .iter()
-        .flat_map(|(&id, accesses)| {
-            accesses
-                .iter()
-                .map(move |access| VarAtomBase::new(id, access.lsb, access.msb))
-        })
-        .collect();
-
-    // 2. Symbolic Execution: Evaluate statements sequentially to update the symbolic state.
-    let effect_initial_store =
-        statements_contain_runtime_effect(module, &decl.statements).then(|| current_store.clone());
-    let (final_store, boundaries) = recover_unrolled::eval_statements(
-        module,
-        current_store,
-        BoundaryMap::default(),
-        &decl.statements,
-        arena,
-        loop_candidates,
-        None,
-    )?;
-    let mut effects = CombEffectCollector::default();
-    if let Some(effect_initial_store) = effect_initial_store {
-        collect_comb_effects_statements(
-            module,
-            effect_initial_store,
-            &decl.statements,
-            arena,
-            &mut effects,
-        )?;
-    }
-
-    // 3. Path Extraction: Convert the final symbolic store into a list of LogicPaths.
-    // Each LogicPath represents a modified bit-range and the logic required to compute it.
-    let mut paths = Vec::new();
-    for (id, range_store) in &final_store {
-        if module.variables[id].affiliation == veryl_analyzer::symbol::Affiliation::AlwaysComb {
-            continue;
-        }
-        for (&lsb, (val_opt, width, origin)) in &range_store.ranges {
-            if let Some((expr, sources)) = val_opt {
-                let msb = lsb + width - 1;
-
-                // Calculate relative bit positions by adjusting for the range's origin.
-                let rel_lsb = lsb - origin;
-                let rel_msb = msb - origin;
-                let original_width = get_width(*expr, arena);
-
-                // If not using the entire stored node, apply Slice
-                let final_expr = if rel_lsb == 0 && *width == original_width {
-                    *expr
-                } else {
-                    arena.alloc(SLTNode::Slice {
-                        expr: *expr,
-                        access: BitAccess::new(rel_lsb, rel_msb),
-                    })?
-                };
-
-                paths.push(LogicPath::<VarId> {
-                    target: LogicPathTarget::Var(VarAtomBase::new(*id, lsb, msb)),
-                    sources: sources.clone(),
-                    previous_sources: sources
-                        .iter()
-                        .copied()
-                        .filter(|source| {
-                            source.id != *id || !source.access.overlaps(&BitAccess::new(lsb, msb))
-                        })
-                        .filter(|source| {
-                            written_atoms.iter().any(|written| {
-                                written.id == source.id && written.access.overlaps(&source.access)
-                            })
-                        })
-                        .collect(),
-                    address_sources: HashSet::default(),
-                    local_inputs: Vec::new(),
-                    order_before: HashSet::default(),
-                    comb_capture_enable_sites: Vec::new(),
-                    pre_lower_nodes: Vec::new(),
-                    expr: final_expr,
-                });
-            }
-        }
-    }
-    let mut process_sensitivity = effects.sensitivity;
-    for path in &paths {
-        process_sensitivity.extend(path.sources.iter().copied());
-    }
-    let process_sensitivity = subtract_written_sensitivity(process_sensitivity, &written_atoms);
-    let process_sensitivity: Vec<_> = process_sensitivity.into_iter().collect();
-    for observer in &mut effects.observers {
-        observer.sensitivity = process_sensitivity.clone();
-        observer.written_input_atoms = observer
-            .observed_inputs
-            .iter()
-            .chain(observer.position_inputs.iter())
-            .copied()
-            .filter(|atom| {
-                written_atoms
-                    .iter()
-                    .any(|written| written.id == atom.id && written.access.overlaps(&atom.access))
-            })
-            .collect();
-        let mut written_inputs = HashSet::default();
-        for atom in &observer.written_input_atoms {
-            written_inputs.insert(atom.id);
-        }
-        observer.written_inputs = written_inputs.into_iter().collect();
-    }
-    dump_comb_path_stats_if_requested(module, &paths, arena);
-    Ok((
-        paths,
-        final_store,
-        boundaries,
-        effects.observers,
-        effects.sites,
-    ))
-}
-
-#[derive(Default)]
-struct CombPathStats {
-    nodes: usize,
-    for_folds: usize,
-    muxes: usize,
-    inputs: usize,
-}
-
-fn dump_comb_path_stats_if_requested(
-    module: &Module,
-    paths: &[LogicPath<VarId>],
-    arena: &SLTNodeArena<VarId>,
-) {
-    if std::env::var_os("CELOX_COMB_PATH_STATS").is_none() {
-        return;
-    }
-
-    let module_name = resource_table::get_str_value(module.name).unwrap_or_default();
-    if let Some(filter) = std::env::var_os("CELOX_COMB_PATH_MODULE")
-        && !module_name.contains(filter.to_string_lossy().as_ref())
-    {
-        return;
-    }
-
-    let mut entries = Vec::new();
-    let mut total_nodes = 0usize;
-    let mut total_for_folds = 0usize;
-    let mut total_muxes = 0usize;
-    let mut total_inputs = 0usize;
-    for path in paths {
-        let mut visited = HashSet::default();
-        let mut stats = CombPathStats::default();
-        collect_comb_path_stats(path.expr, arena, &mut visited, &mut stats);
-        total_nodes += stats.nodes;
-        total_for_folds += stats.for_folds;
-        total_muxes += stats.muxes;
-        total_inputs += stats.inputs;
-        let target = match &path.target {
-            LogicPathTarget::Var(var) => module.variables.get(&var.id).map_or_else(
-                || var.to_string(),
-                |info| format!("{}[{}:{}]", info.path, var.access.msb, var.access.lsb),
-            ),
-            LogicPathTarget::CombCaptureEvent { site_id, .. } => {
-                format!("capture_event({site_id})")
-            }
-        };
-        entries.push((
-            stats.nodes,
-            stats.for_folds,
-            stats.muxes,
-            stats.inputs,
-            target,
-        ));
-    }
-    entries.sort_by(|a, b| b.cmp(a));
-
-    eprintln!(
-        "[comb-path-summary] module={} paths={} total_nodes={} total_for_folds={} total_muxes={} total_inputs={}",
-        module_name,
-        paths.len(),
-        total_nodes,
-        total_for_folds,
-        total_muxes,
-        total_inputs,
-    );
-
-    let limit = std::env::var("CELOX_COMB_PATH_STATS_LIMIT")
-        .ok()
-        .and_then(|raw| raw.parse::<usize>().ok())
-        .unwrap_or(20);
-    for (rank, (nodes, for_folds, muxes, inputs, target)) in
-        entries.into_iter().take(limit).enumerate()
-    {
-        eprintln!(
-            "[comb-path-stats] module={} rank={} target={} nodes={} for_folds={} muxes={} inputs={}",
-            module_name,
-            rank + 1,
-            target,
-            nodes,
-            for_folds,
-            muxes,
-            inputs,
-        );
-    }
-}
-
-fn collect_comb_path_stats(
-    node: NodeId,
-    arena: &SLTNodeArena<VarId>,
-    visited: &mut HashSet<NodeId>,
-    stats: &mut CombPathStats,
-) {
-    if !visited.insert(node) {
-        return;
-    }
-    stats.nodes += 1;
-    match arena.get(node) {
-        SLTNode::Input { .. } => stats.inputs += 1,
-        SLTNode::Constant(_, _, _, _) => {}
-        SLTNode::Binary(lhs, _, rhs) => {
-            collect_comb_path_stats(*lhs, arena, visited, stats);
-            collect_comb_path_stats(*rhs, arena, visited, stats);
-        }
-        SLTNode::Unary(_, inner) => collect_comb_path_stats(*inner, arena, visited, stats),
-        SLTNode::Mux {
-            cond,
-            then_expr,
-            else_expr,
-        } => {
-            stats.muxes += 1;
-            collect_comb_path_stats(*cond, arena, visited, stats);
-            collect_comb_path_stats(*then_expr, arena, visited, stats);
-            collect_comb_path_stats(*else_expr, arena, visited, stats);
-        }
-        SLTNode::Concat(parts) => {
-            for (part, _) in parts {
-                collect_comb_path_stats(*part, arena, visited, stats);
-            }
-        }
-        SLTNode::Slice { expr, .. } => collect_comb_path_stats(*expr, arena, visited, stats),
-        SLTNode::ForFold {
-            start,
-            end,
-            initials,
-            updates,
-            effects,
-            continue_cond,
-            ..
-        } => {
-            stats.for_folds += 1;
-            if let SLTLoopBound::Expr(node) = start {
-                collect_comb_path_stats(*node, arena, visited, stats);
-            }
-            if let SLTLoopBound::Expr(node) = end {
-                collect_comb_path_stats(*node, arena, visited, stats);
-            }
-            for init in initials {
-                collect_comb_path_stats(init.expr, arena, visited, stats);
-            }
-            for update in updates {
-                collect_comb_path_stats(update.expr, arena, visited, stats);
-            }
-            for effect in effects {
-                if let Some(guard) = effect.guard {
-                    collect_comb_path_stats(guard, arena, visited, stats);
-                }
-                for arg in &effect.args {
-                    collect_comb_path_stats(*arg, arena, visited, stats);
-                }
-            }
-            collect_comb_path_stats(*continue_cond, arena, visited, stats);
-        }
-        SLTNode::ForFoldGroup {
-            entry_guard,
-            states,
-            ..
-        } => {
-            stats.for_folds += 1;
-            collect_comb_path_stats(*entry_guard, arena, visited, stats);
-            for state in states {
-                collect_comb_path_stats(state.initial, arena, visited, stats);
-                collect_comb_path_stats(state.update, arena, visited, stats);
-            }
-        }
-    }
-}
-
-fn const_for_bound_i64(bound: &ForBound) -> Option<i64> {
-    match bound {
-        ForBound::Const(v) => (*v).try_into().ok(),
-        ForBound::Expression(expr) => eval_constexpr(expr)?.to_i64(),
-    }
-}
-
-fn eval_statement(
-    module: &Module,
-    store: SymbolicStore<VarId>,
-    boundaries: HashMap<VarId, BTreeSet<usize>>,
-    stmt: &Statement,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(SymbolicStore<VarId>, HashMap<VarId, BTreeSet<usize>>), ParserError> {
-    match stmt {
-        Statement::Assign(assign) => eval_assign(module, store, boundaries, assign, arena),
-        Statement::If(if_stmt) => eval_if(module, store, boundaries, if_stmt, arena),
-        Statement::Case(case_stmt) => eval_case(module, store, boundaries, case_stmt, arena),
-        Statement::For(for_stmt) => eval_for(module, store, boundaries, for_stmt, arena),
-        Statement::IfReset(ir) => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "if_reset".to_string(),
-            Some(&ir.token),
-        )),
-        Statement::SystemFunctionCall(_) => Ok((store, boundaries)),
-        Statement::FunctionCall(fc) => eval_statement_form_function_call(
-            module,
-            store,
-            boundaries,
-            fc,
-            arena,
-            LoweringPhase::CombLowering,
-        ),
-        Statement::TbMethodCall(_) => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "testbench method call".to_string(),
-            None,
-        )),
-        Statement::Break => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "break".to_string(),
-            None,
-        )),
-        Statement::Unsupported(_) => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "unsupported statement".to_string(),
-            None,
-        )),
-        Statement::Null => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "null".to_string(),
-            None,
-        )),
-    }
-}
-
-fn eval_statement_with_recovery(
-    module: &Module,
-    store: SymbolicStore<VarId>,
-    boundaries: HashMap<VarId, BTreeSet<usize>>,
-    stmt: &Statement,
-    arena: &mut SLTNodeArena<VarId>,
-    loop_candidates: &[LoopRecoveryCandidate],
-    active_guard: Option<&ActiveGuard>,
-) -> Result<(SymbolicStore<VarId>, HashMap<VarId, BTreeSet<usize>>), ParserError> {
-    match stmt {
-        Statement::If(if_stmt) => eval_if_with_recovery(
-            module,
-            store,
-            boundaries,
-            if_stmt,
-            arena,
-            loop_candidates,
-            active_guard,
-        ),
-        Statement::Case(case_stmt) => eval_case_with_recovery(
-            module,
-            store,
-            boundaries,
-            case_stmt,
-            arena,
-            loop_candidates,
-            active_guard,
-        ),
-        _ => eval_statement(module, store, boundaries, stmt, arena),
-    }
-}
-
-fn eval_statements(
-    module: &Module,
-    store: SymbolicStore<VarId>,
-    boundaries: BoundaryMap<VarId>,
-    statements: &[Statement],
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-    statements
-        .iter()
-        .try_fold((store, boundaries), |(store, boundaries), stmt| {
-            eval_statement(module, store, boundaries, stmt, arena)
-        })
-}
-
-fn eval_statements_with_recovery(
-    module: &Module,
-    store: SymbolicStore<VarId>,
-    boundaries: BoundaryMap<VarId>,
-    statements: &[Statement],
-    arena: &mut SLTNodeArena<VarId>,
-    loop_candidates: &[LoopRecoveryCandidate],
-    active_guard: Option<&ActiveGuard>,
-) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-    recover_unrolled::eval_statements(
-        module,
-        store,
-        boundaries,
-        statements,
-        arena,
-        loop_candidates,
-        active_guard,
-    )
-}
-
-fn eval_case_with_recovery(
-    module: &Module,
-    store: SymbolicStore<VarId>,
-    boundaries: BoundaryMap<VarId>,
-    case_stmt: &CaseStatement,
-    arena: &mut SLTNodeArena<VarId>,
-    loop_candidates: &[LoopRecoveryCandidate],
-    active_guard: Option<&ActiveGuard>,
-) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-    fn eval_from_arm(
-        module: &Module,
-        store: SymbolicStore<VarId>,
-        boundaries: BoundaryMap<VarId>,
-        case_stmt: &CaseStatement,
-        arm_index: usize,
-        arena: &mut SLTNodeArena<VarId>,
-        loop_candidates: &[LoopRecoveryCandidate],
-        active_guard: Option<&ActiveGuard>,
-    ) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-        let Some(arm) = case_stmt.arms.get(arm_index) else {
-            return eval_statements_with_recovery(
-                module,
-                store,
-                boundaries,
-                &case_stmt.default,
-                arena,
-                loop_candidates,
-                active_guard,
-            );
-        };
-
-        let cond = case_arm_condition_expr(&case_stmt.case_target, &arm.patterns);
-        let ((cond_expr, cond_sources), cond_bounds) =
-            eval_expression(module, &store, &cond, arena, None)?;
-        let cond_expr = procedural_condition(arena, cond_expr)?;
-        let boundaries = merge_boundaries(boundaries, cond_bounds);
-
-        if let Some(cond_val) = constant_bool(arena, cond_expr) {
-            return if cond_val {
-                eval_statements_with_recovery(
-                    module,
-                    store,
-                    boundaries,
-                    &arm.body,
-                    arena,
-                    loop_candidates,
-                    active_guard,
-                )
-            } else {
-                eval_from_arm(
-                    module,
-                    store,
-                    boundaries,
-                    case_stmt,
-                    arm_index + 1,
-                    arena,
-                    loop_candidates,
-                    active_guard,
-                )
-            };
-        }
-
-        let then_guard = combine_active_guard(arena, active_guard, cond_expr, &cond_sources)?;
-        let false_condition = invert_active_condition(arena, cond_expr)?;
-        let else_guard = combine_active_guard(arena, active_guard, false_condition, &cond_sources)?;
-
-        let (then_store, then_boundaries) = eval_statements_with_recovery(
-            module,
-            store.clone(),
-            boundaries.clone(),
-            &arm.body,
-            arena,
-            loop_candidates,
-            Some(&then_guard),
-        )?;
-        let (else_store, else_boundaries) = eval_from_arm(
-            module,
-            store,
-            boundaries,
-            case_stmt,
-            arm_index + 1,
-            arena,
-            loop_candidates,
-            Some(&else_guard),
-        )?;
-
-        Ok((
-            merge_symbolic_stores(
-                module,
-                &then_store,
-                &else_store,
-                cond_expr,
-                &cond_sources,
-                arena,
-            )?,
-            merge_boundaries(then_boundaries, else_boundaries),
-        ))
-    }
-
-    eval_from_arm(
-        module,
-        store,
-        boundaries,
-        case_stmt,
-        0,
-        arena,
-        loop_candidates,
-        active_guard,
-    )
-}
-
-fn eval_case(
-    module: &Module,
-    store: SymbolicStore<VarId>,
-    boundaries: BoundaryMap<VarId>,
-    case_stmt: &CaseStatement,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-    fn eval_from_arm(
-        module: &Module,
-        store: SymbolicStore<VarId>,
-        boundaries: BoundaryMap<VarId>,
-        case_stmt: &CaseStatement,
-        arm_index: usize,
-        arena: &mut SLTNodeArena<VarId>,
-    ) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-        let Some(arm) = case_stmt.arms.get(arm_index) else {
-            return eval_statements(module, store, boundaries, &case_stmt.default, arena);
-        };
-
-        let cond = case_arm_condition_expr(&case_stmt.case_target, &arm.patterns);
-        let ((cond_expr, cond_sources), cond_bounds) =
-            eval_expression(module, &store, &cond, arena, None)?;
-        let cond_expr = procedural_condition(arena, cond_expr)?;
-        let boundaries = merge_boundaries(boundaries, cond_bounds);
-
-        if let Some(cond_val) = constant_bool(arena, cond_expr) {
-            return if cond_val {
-                eval_statements(module, store, boundaries, &arm.body, arena)
-            } else {
-                eval_from_arm(module, store, boundaries, case_stmt, arm_index + 1, arena)
-            };
-        }
-
-        let (then_store, then_boundaries) =
-            eval_statements(module, store.clone(), boundaries.clone(), &arm.body, arena)?;
-        let (else_store, else_boundaries) =
-            eval_from_arm(module, store, boundaries, case_stmt, arm_index + 1, arena)?;
-
-        Ok((
-            merge_symbolic_stores(
-                module,
-                &then_store,
-                &else_store,
-                cond_expr,
-                &cond_sources,
-                arena,
-            )?,
-            merge_boundaries(then_boundaries, else_boundaries),
-        ))
-    }
-
-    eval_from_arm(module, store, boundaries, case_stmt, 0, arena)
-}
-
-fn bool_node(arena: &mut SLTNodeArena<VarId>, value: bool) -> Result<NodeId, SLTNodeFactsError> {
-    arena.alloc(SLTNode::Constant(
-        BigUint::from(value as u8),
-        BigUint::from(0u8),
-        1,
-        false,
-    ))
-}
-
-fn function_assigns_whole_var(assign: &AssignStatement, var_id: VarId) -> bool {
-    assign.dst.len() == 1
-        && assign.dst[0].id == var_id
-        && assign.dst[0].index.0.is_empty()
-        && assign.dst[0].select.0.is_empty()
-        && assign.dst[0].select.1.is_none()
-}
-
-fn constant_bool(arena: &SLTNodeArena<VarId>, node: NodeId) -> Option<bool> {
-    match arena.get(node) {
-        SLTNode::Constant(val, _, _, _) => Some(*val != BigUint::from(0u8)),
-        _ => None,
-    }
-}
-
-/// Convert an expression result to the boolean used by procedural control.
-///
-/// Unlike the conditional operator, an unknown condition in an `if`, `case`,
-/// or loop does not merge both paths: only a definite one takes the true
-/// path.  Keep that distinction in the SLT instead of relying on a backend's
-/// treatment of a four-state mux.
-fn procedural_condition<A: Clone + Eq + Hash>(
-    arena: &mut SLTNodeArena<A>,
-    condition: NodeId,
-) -> Result<NodeId, SLTNodeFactsError> {
-    if let SLTNode::Constant(value, unknown, width, _) = arena.get(condition) {
-        let known_mask = if *width == 0 {
-            BigUint::from(0u8)
-        } else {
-            let width_mask = (BigUint::from(1u8) << *width) - BigUint::from(1u8);
-            &width_mask ^ (unknown & &width_mask)
-        };
-        return arena.alloc(SLTNode::Constant(
-            BigUint::from(u8::from((value & known_mask) != BigUint::from(0u8))),
-            BigUint::from(0u8),
-            1,
-            false,
-        ));
-    }
-    let truth = arena.alloc(SLTNode::Unary(UnaryOp::Or, condition))?;
-    arena.alloc(SLTNode::Unary(UnaryOp::ToTwoState, truth))
-}
-
-fn combine_active_guard(
-    arena: &mut SLTNodeArena<VarId>,
-    outer: Option<&ActiveGuard>,
-    condition: NodeId,
-    condition_sources: &HashSet<VarAtomBase<VarId>>,
-) -> Result<ActiveGuard, SLTNodeFactsError> {
-    let Some((outer_expr, outer_sources)) = outer else {
-        return Ok((condition, condition_sources.clone()));
-    };
-    let expr = arena.alloc(SLTNode::Binary(*outer_expr, BinaryOp::LogicAnd, condition))?;
-    let mut sources = outer_sources.clone();
-    sources.extend(condition_sources.iter().copied());
-    Ok((expr, sources))
-}
-
-fn invert_active_condition(
-    arena: &mut SLTNodeArena<VarId>,
-    condition: NodeId,
-) -> Result<NodeId, SLTNodeFactsError> {
-    arena.alloc(SLTNode::Unary(UnaryOp::LogicNot, condition))
-}
-
-fn merge_control_expr(
-    cond_expr: NodeId,
-    then_expr: NodeId,
-    else_expr: NodeId,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<NodeId, SLTNodeFactsError> {
-    if then_expr == else_expr {
-        Ok(then_expr)
-    } else {
-        arena.alloc(SLTNode::Mux {
-            cond: cond_expr,
-            then_expr,
-            else_expr,
-        })
-    }
-}
-
-fn merge_symbolic_stores(
-    module: &Module,
-    then_store: &SymbolicStore<VarId>,
-    else_store: &SymbolicStore<VarId>,
-    cond_expr: NodeId,
-    cond_sources: &HashSet<VarAtomBase<VarId>>,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<SymbolicStore<VarId>, ParserError> {
-    let mut merged_store = SymbolicStore::default();
-    for id in then_store.keys() {
-        let t_range_store = &then_store[id];
-        let e_range_store = &else_store[id];
-
-        let mut merged_range_store = RangeStore {
-            ranges: std::collections::BTreeMap::new(),
-        };
-
-        let mut all_lsbs: BTreeSet<usize> = t_range_store.ranges.keys().cloned().collect();
-        all_lsbs.extend(e_range_store.ranges.keys().cloned());
-
-        let var = &module.variables[id];
-        let var_width = resolve_total_width(module, var)?;
-        let mut lsbs_vec: Vec<usize> = all_lsbs.into_iter().collect();
-        lsbs_vec.push(var_width);
-
-        for i in 0..lsbs_vec.len() - 1 {
-            let lsb = lsbs_vec[i];
-            let next_lsb = lsbs_vec[i + 1];
-            let access = BitAccess::new(lsb, next_lsb - 1);
-
-            let then_parts = t_range_store
-                .get_parts(access)
-                .map_err(|error| range_store_error("conditional merge", error, None))?;
-            let else_parts = e_range_store
-                .get_parts(access)
-                .map_err(|error| range_store_error("conditional merge", error, None))?;
-            let (t_expr, t_sources) =
-                combine_parts_with_default(*id, lsb, then_parts.clone(), arena)?;
-            let (e_expr, e_sources) =
-                combine_parts_with_default(*id, lsb, else_parts.clone(), arena)?;
-
-            let t_modified = then_parts.iter().any(|(v, _)| v.is_some());
-            let e_modified = else_parts.iter().any(|(v, _)| v.is_some());
-
-            let result_val = if !t_modified && !e_modified {
-                None
-            } else if t_expr == e_expr {
-                let mut sources = t_sources;
-                sources.extend(e_sources);
-                Some((t_expr, sources))
-            } else {
-                let mut sources = cond_sources.clone();
-                sources.extend(t_sources);
-                sources.extend(e_sources);
-
-                Some((
-                    arena.alloc(SLTNode::Mux {
-                        cond: cond_expr,
-                        then_expr: t_expr,
-                        else_expr: e_expr,
-                    })?,
-                    sources,
-                ))
-            };
-
-            merged_range_store
-                .ranges
-                .insert(lsb, (result_val, next_lsb - lsb, lsb));
-        }
-
-        merged_store.insert(*id, merged_range_store);
-    }
-
-    Ok(merged_store)
-}
-
-fn apply_loop_continue_guard(
-    module: &Module,
-    state: LoopControlState,
-    next_store: SymbolicStore<VarId>,
-    next_boundaries: BoundaryMap<VarId>,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<LoopControlState, ParserError> {
-    let base_store = state.store.clone();
-    let boundaries = merge_boundaries(state.boundaries, next_boundaries);
-
-    if matches!(constant_bool(arena, state.continue_expr), Some(true)) {
-        Ok(LoopControlState {
-            store: next_store,
-            boundaries,
-            ..state
-        })
-    } else {
-        let merged_store = merge_symbolic_stores(
-            module,
-            &next_store,
-            &base_store,
-            state.continue_expr,
-            &state.continue_sources,
-            arena,
-        )?;
-        Ok(LoopControlState {
-            store: merged_store,
-            boundaries,
-            ..state
-        })
-    }
-}
-
-fn statement_contains_break(stmt: &Statement) -> bool {
-    match stmt {
-        Statement::Break => true,
-        Statement::If(if_stmt) => {
-            if_stmt.true_side.iter().any(statement_contains_break)
-                || if_stmt.false_side.iter().any(statement_contains_break)
-        }
-        Statement::Case(case_stmt) => {
-            case_stmt
-                .arms
-                .iter()
-                .any(|arm| arm.body.iter().any(statement_contains_break))
-                || case_stmt.default.iter().any(statement_contains_break)
-        }
-        Statement::For(for_stmt) => for_stmt.body.iter().any(statement_contains_break),
-        Statement::IfReset(if_reset) => {
-            if_reset.true_side.iter().any(statement_contains_break)
-                || if_reset.false_side.iter().any(statement_contains_break)
-        }
-        Statement::Assign(_)
-        | Statement::SystemFunctionCall(_)
-        | Statement::FunctionCall(_)
-        | Statement::TbMethodCall(_)
-        | Statement::Unsupported(_)
-        | Statement::Null => false,
-    }
-}
-
-fn eval_loop_statement(
-    module: &Module,
-    state: LoopControlState,
-    stmt: &Statement,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<LoopControlState, ParserError> {
-    if matches!(constant_bool(arena, state.continue_expr), Some(false)) {
-        return Ok(state);
-    }
-
-    match stmt {
-        Statement::Assign(assign) => {
-            let guard_state = state.clone();
-            let (next_store, next_boundaries) =
-                eval_assign(module, state.store, state.boundaries, assign, arena)?;
-            apply_loop_continue_guard(module, guard_state, next_store, next_boundaries, arena)
-        }
-        Statement::If(if_stmt) => {
-            if statement_contains_break(stmt) {
-                eval_loop_if(module, state, if_stmt, arena)
-            } else {
-                let guard_state = state.clone();
-                let (next_store, next_boundaries) =
-                    eval_if(module, state.store, state.boundaries, if_stmt, arena)?;
-                apply_loop_continue_guard(module, guard_state, next_store, next_boundaries, arena)
-            }
-        }
-        Statement::Case(case_stmt) => eval_loop_case(module, state, case_stmt, arena),
-        Statement::For(for_stmt) => {
-            let guard_state = state.clone();
-            let (next_store, next_boundaries) =
-                eval_for(module, state.store, state.boundaries, for_stmt, arena)?;
-            apply_loop_continue_guard(module, guard_state, next_store, next_boundaries, arena)
-        }
-        Statement::Break => Ok(LoopControlState {
-            continue_expr: bool_node(arena, false)?,
-            continue_sources: HashSet::default(),
-            ..state
-        }),
-        Statement::IfReset(ir) => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "if_reset".to_string(),
-            Some(&ir.token),
-        )),
-        Statement::SystemFunctionCall(_) => Ok(state),
-        Statement::FunctionCall(fc) => {
-            let guard_state = state.clone();
-            let (next_store, next_boundaries) = eval_statement_form_function_call(
-                module,
-                state.store,
-                state.boundaries,
-                fc,
-                arena,
-                LoweringPhase::CombLowering,
-            )?;
-            apply_loop_continue_guard(module, guard_state, next_store, next_boundaries, arena)
-        }
-        Statement::TbMethodCall(_) => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "testbench method call".to_string(),
-            None,
-        )),
-        Statement::Unsupported(_) => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "unsupported statement".to_string(),
-            None,
-        )),
-        Statement::Null => Err(ParserError::illegal_context(
-            "statement in always_comb",
-            "null".to_string(),
-            None,
-        )),
-    }
-}
-
-fn eval_loop_case(
-    module: &Module,
-    state: LoopControlState,
-    case_stmt: &CaseStatement,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<LoopControlState, ParserError> {
-    fn eval_from_arm(
-        module: &Module,
-        state: LoopControlState,
-        case_stmt: &CaseStatement,
-        arm_index: usize,
-        arena: &mut SLTNodeArena<VarId>,
-    ) -> Result<LoopControlState, ParserError> {
-        let Some(arm) = case_stmt.arms.get(arm_index) else {
-            return case_stmt
-                .default
-                .iter()
-                .try_fold(state, |s, step| eval_loop_statement(module, s, step, arena));
-        };
-
-        let ((cond_expr, cond_sources), cond_bounds) = eval_expression(
-            module,
-            &state.store,
-            &case_arm_condition_expr(&case_stmt.case_target, &arm.patterns),
-            arena,
-            None,
-        )?;
-        let cond_expr = procedural_condition(arena, cond_expr)?;
-        let boundaries = merge_boundaries(state.boundaries, cond_bounds);
-
-        if let Some(cond_val) = constant_bool(arena, cond_expr) {
-            let state = LoopControlState {
-                boundaries,
-                ..state
-            };
-            return if cond_val {
-                arm.body
-                    .iter()
-                    .try_fold(state, |s, step| eval_loop_statement(module, s, step, arena))
-            } else {
-                eval_from_arm(module, state, case_stmt, arm_index + 1, arena)
-            };
-        }
-
-        let then_state = arm.body.iter().try_fold(
-            LoopControlState {
-                store: state.store.clone(),
-                boundaries: boundaries.clone(),
-                continue_expr: state.continue_expr,
-                continue_sources: state.continue_sources.clone(),
-            },
-            |s, step| eval_loop_statement(module, s, step, arena),
-        )?;
-        let else_state = eval_from_arm(
-            module,
-            LoopControlState {
-                store: state.store,
-                boundaries,
-                continue_expr: state.continue_expr,
-                continue_sources: state.continue_sources,
-            },
-            case_stmt,
-            arm_index + 1,
-            arena,
-        )?;
-
-        let mut merged_sources = cond_sources;
-        merged_sources.extend(then_state.continue_sources);
-        merged_sources.extend(else_state.continue_sources);
-
-        Ok(LoopControlState {
-            store: merge_symbolic_stores(
-                module,
-                &then_state.store,
-                &else_state.store,
-                cond_expr,
-                &merged_sources,
-                arena,
-            )?,
-            boundaries: merge_boundaries(then_state.boundaries, else_state.boundaries),
-            continue_expr: merge_control_expr(
-                cond_expr,
-                then_state.continue_expr,
-                else_state.continue_expr,
-                arena,
-            )?,
-            continue_sources: merged_sources,
-        })
-    }
-
-    eval_from_arm(module, state, case_stmt, 0, arena)
-}
-
-fn eval_loop_if(
-    module: &Module,
-    state: LoopControlState,
-    stmt: &IfStatement,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<LoopControlState, ParserError> {
-    let ((cond_expr, cond_sources), cond_bounds) =
-        eval_expression(module, &state.store, &stmt.cond, arena, None)?;
-    let cond_expr = procedural_condition(arena, cond_expr)?;
-    let boundaries = merge_boundaries(state.boundaries, cond_bounds);
-
-    if let Some(cond_val) = constant_bool(arena, cond_expr) {
-        let side = if cond_val {
-            &stmt.true_side
-        } else {
-            &stmt.false_side
-        };
-        return side.iter().try_fold(
-            LoopControlState {
-                boundaries,
-                ..state
-            },
-            |s, step| eval_loop_statement(module, s, step, arena),
-        );
-    }
-
-    let then_state = stmt.true_side.iter().try_fold(
-        LoopControlState {
-            store: state.store.clone(),
-            boundaries: boundaries.clone(),
-            continue_expr: state.continue_expr,
-            continue_sources: state.continue_sources.clone(),
-        },
-        |s, step| eval_loop_statement(module, s, step, arena),
-    )?;
-    let else_state = stmt.false_side.iter().try_fold(
-        LoopControlState {
-            store: state.store,
-            boundaries,
-            continue_expr: state.continue_expr,
-            continue_sources: state.continue_sources,
-        },
-        |s, step| eval_loop_statement(module, s, step, arena),
-    )?;
-
-    let mut merged_sources = cond_sources;
-    merged_sources.extend(then_state.continue_sources);
-    merged_sources.extend(else_state.continue_sources);
-
-    Ok(LoopControlState {
-        store: merge_symbolic_stores(
-            module,
-            &then_state.store,
-            &else_state.store,
-            cond_expr,
-            &merged_sources,
-            arena,
-        )?,
-        boundaries: merge_boundaries(then_state.boundaries, else_state.boundaries),
-        continue_expr: merge_control_expr(
-            cond_expr,
-            then_state.continue_expr,
-            else_state.continue_expr,
-            arena,
-        )?,
-        continue_sources: merged_sources,
-    })
-}
-
-fn extract_store_updates(
-    store_before: &SymbolicStore<VarId>,
-    store_after: &SymbolicStore<VarId>,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<Vec<(VarAtomBase<VarId>, NodeId, HashSet<VarAtomBase<VarId>>)>, SLTNodeFactsError> {
-    let mut updates = Vec::new();
-
-    for (id, range_store_after) in store_after {
-        let Some(range_store_before) = store_before.get(id) else {
-            continue;
-        };
-
-        for (&lsb, (val_opt, width, origin)) in &range_store_after.ranges {
-            if range_store_before.ranges.get(&lsb) == Some(&(val_opt.clone(), *width, *origin)) {
-                continue;
-            }
-
-            let Some((expr, sources)) = val_opt else {
-                continue;
-            };
-
-            let msb = lsb + width - 1;
-            let rel_lsb = lsb - origin;
-            let rel_msb = msb - origin;
-            let original_width = get_width(*expr, arena);
-            let final_expr = if rel_lsb == 0 && *width == original_width {
-                *expr
-            } else {
-                arena.alloc(SLTNode::Slice {
-                    expr: *expr,
-                    access: BitAccess::new(rel_lsb, rel_msb),
-                })?
-            };
-
-            updates.push((VarAtomBase::new(*id, lsb, msb), final_expr, sources.clone()));
-        }
-    }
-
-    Ok(updates)
-}
-
-fn eval_for_bound(
-    module: &Module,
-    store: &SymbolicStore<VarId>,
-    bound: &ForBound,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<
-    (
-        SLTLoopBound,
-        HashSet<VarAtomBase<VarId>>,
-        BoundaryMap<VarId>,
-    ),
-    ParserError,
-> {
-    match bound {
-        ForBound::Const(v) => Ok((
-            SLTLoopBound::Const(*v),
-            HashSet::default(),
-            BoundaryMap::default(),
-        )),
-        ForBound::Expression(expr) => {
-            let ((node, sources), bounds) = eval_expression(module, store, expr, arena, None)?;
-            Ok((SLTLoopBound::Expr(node), sources, bounds))
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum LoopBoundStatus {
-    FitsLoopType,
-    ExclusiveUpperSentinel,
-    OutOfRange,
-}
-
-fn loop_bound_status(bound: &ForBound, width: usize, signed: bool) -> Option<LoopBoundStatus> {
-    let value = match bound {
-        ForBound::Const(v) => BigInt::from(*v),
-        ForBound::Expression(expr) => {
-            if !expr.comptime().is_const {
-                return None;
-            }
-            let value = expr.comptime().get_value().ok()?;
-            match value {
-                Value::U64(v) => {
-                    if v.signed {
-                        BigInt::from(v.to_i64()?)
-                    } else {
-                        BigInt::from(v.to_u64()?)
-                    }
-                }
-                Value::BigUint(v) => {
-                    if v.signed {
-                        v.to_bigint()?
-                    } else {
-                        BigInt::from_biguint(Sign::Plus, (*v.payload).clone())
-                    }
-                }
-            }
-        }
-    };
-
-    if signed {
-        let max = (BigInt::from(1u8) << (width.saturating_sub(1))) - BigInt::from(1u8);
-        let min = -(BigInt::from(1u8) << (width.saturating_sub(1)));
-        Some(if value >= min && value <= max {
-            LoopBoundStatus::FitsLoopType
-        } else if value == max + BigInt::from(1u8) {
-            LoopBoundStatus::ExclusiveUpperSentinel
-        } else {
-            LoopBoundStatus::OutOfRange
-        })
-    } else {
-        let max = (BigUint::from(1u8) << width) - BigUint::from(1u8);
-        let max = BigInt::from_biguint(Sign::Plus, max);
-        Some(if value.sign() != Sign::Minus && value <= max {
-            LoopBoundStatus::FitsLoopType
-        } else if value == max + BigInt::from(1u8) {
-            LoopBoundStatus::ExclusiveUpperSentinel
-        } else {
-            LoopBoundStatus::OutOfRange
-        })
-    }
-}
-
-fn inclusive_of(range: &ForRange) -> bool {
-    match range {
-        ForRange::Forward { inclusive, .. }
-        | ForRange::Reverse { inclusive, .. }
-        | ForRange::Stepped { inclusive, .. } => *inclusive,
-    }
-}
-
-fn collect_written_accesses(
-    module: &Module,
-    statements: &[Statement],
-    out: &mut HashMap<VarId, Vec<BitAccess>>,
-) -> Result<(), ParserError> {
-    for stmt in statements {
-        match stmt {
-            Statement::Assign(assign) => {
-                for dst in &assign.dst {
-                    collect_written_destination(module, out, dst)?;
-                }
-            }
-            Statement::If(if_stmt) => {
-                collect_written_accesses(module, &if_stmt.true_side, out)?;
-                collect_written_accesses(module, &if_stmt.false_side, out)?;
-            }
-            Statement::Case(case_stmt) => {
-                for arm in &case_stmt.arms {
-                    collect_written_accesses(module, &arm.body, out)?;
-                }
-                collect_written_accesses(module, &case_stmt.default, out)?;
-            }
-            Statement::For(for_stmt) => collect_written_accesses(module, &for_stmt.body, out)?,
-            Statement::IfReset(if_reset) => {
-                collect_written_accesses(module, &if_reset.true_side, out)?;
-                collect_written_accesses(module, &if_reset.false_side, out)?;
-            }
-            Statement::FunctionCall(call) => {
-                for dsts in call.outputs.values() {
-                    for dst in dsts {
-                        collect_written_destination(module, out, dst)?;
-                    }
-                }
-            }
-            Statement::SystemFunctionCall(_)
-            | Statement::TbMethodCall(_)
-            | Statement::Break
-            | Statement::Unsupported(_)
-            | Statement::Null => {}
-        }
-    }
-    Ok(())
-}
-
-fn collect_written_destination(
-    module: &Module,
-    out: &mut HashMap<VarId, Vec<BitAccess>>,
-    dst: &veryl_analyzer::ir::AssignDestination,
-) -> Result<(), ParserError> {
-    let access = eval_var_select(module, dst.id, &dst.index, &dst.select)?;
-    out.entry(dst.id).or_default().push(access);
-    Ok(())
-}
-
-fn eval_for(
-    module: &Module,
-    store: SymbolicStore<VarId>,
-    boundaries: HashMap<VarId, BTreeSet<usize>>,
-    for_stmt: &ForStatement,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(SymbolicStore<VarId>, HashMap<VarId, BTreeSet<usize>>), ParserError> {
-    eval_for_with_effects(module, store, boundaries, for_stmt, arena, &[])
-        .map(|(store, boundaries, _)| (store, boundaries))
-}
-
-fn eval_for_with_effects(
-    module: &Module,
-    store: SymbolicStore<VarId>,
-    boundaries: HashMap<VarId, BTreeSet<usize>>,
-    for_stmt: &ForStatement,
-    arena: &mut SLTNodeArena<VarId>,
-    effects: &[SLTForEffect],
-) -> Result<
-    (
-        SymbolicStore<VarId>,
-        HashMap<VarId, BTreeSet<usize>>,
-        Option<NodeId>,
-    ),
-    ParserError,
-> {
-    let Some(loop_width) = for_stmt.var_type.total_width() else {
-        return Err(ParserError::unsupported(
-            65,
-            LoweringPhase::CombLowering,
-            "for loop variable width",
-            format!("{:?}", for_stmt.var_name),
-            Some(&for_stmt.token),
-        ));
-    };
-    let (start_bound, end_bound) = match &for_stmt.range {
-        ForRange::Forward { start, end, .. }
-        | ForRange::Reverse { start, end, .. }
-        | ForRange::Stepped { start, end, .. } => (start, end),
-    };
-    let start_status = loop_bound_status(start_bound, loop_width, for_stmt.var_type.signed);
-    let end_status = loop_bound_status(end_bound, loop_width, for_stmt.var_type.signed);
-    // Keep the exclusive upper sentinel used for full-range iteration such as
-    // `0..256` on an 8-bit loop variable, but reject bounds that would
-    // actually force the loop variable outside its representable range.
-    if matches!(
-        start_status,
-        Some(LoopBoundStatus::OutOfRange | LoopBoundStatus::ExclusiveUpperSentinel)
-    ) || matches!(end_status, Some(LoopBoundStatus::OutOfRange))
-        || (inclusive_of(&for_stmt.range)
-            && end_status == Some(LoopBoundStatus::ExclusiveUpperSentinel))
-    {
-        return Err(ParserError::illegal_context(
-            "for loop bound exceeding i32 loop variable",
-            format!("{:?}", for_stmt.var_name),
-            Some(&for_stmt.token),
-        ));
-    }
-
-    let mut symbolic_store = store.clone();
-    let mut written_accesses = HashMap::default();
-    collect_written_accesses(module, &for_stmt.body, &mut written_accesses)?;
-    for (id, accesses) in written_accesses {
-        let width = resolve_total_width(module, &module.variables[&id])?;
-        let mut loop_store = RangeStore::new(None, width);
-        let mut covered = vec![false; width];
-        for access in accesses {
-            for slot in covered.iter_mut().take(access.msb + 1).skip(access.lsb) {
-                *slot = true;
-            }
-        }
-        let original = store
-            .get(&id)
-            .cloned()
-            .unwrap_or_else(|| RangeStore::new(None, width));
-        let mut bit = 0usize;
-        while bit < width {
-            if covered[bit] {
-                bit += 1;
-                continue;
-            }
-            let start = bit;
-            while bit < width && !covered[bit] {
-                bit += 1;
-            }
-            let end = bit - 1;
-            let access = BitAccess::new(start, end);
-            let parts = original
-                .get_parts(access)
-                .map_err(|error| range_store_error("for-loop state", error, None))?;
-            let (expr, sources) = combine_parts_with_default(id, access.lsb, parts, arena)?;
-            loop_store
-                .update(access, Some((expr, sources)))
-                .map_err(|error| range_store_error("for-loop state", error, None))?;
-        }
-        symbolic_store.insert(id, loop_store);
-    }
-    symbolic_store.insert(for_stmt.var_id, RangeStore::new(None, loop_width));
-    let iter_store_before = symbolic_store.clone();
-
-    let loop_state = for_stmt.body.iter().try_fold(
-        LoopControlState {
-            store: symbolic_store,
-            boundaries,
-            continue_expr: bool_node(arena, true)?,
-            continue_sources: HashSet::default(),
-        },
-        |state, stmt| eval_loop_statement(module, state, stmt, arena),
-    )?;
-    let iter_store_after = loop_state.store;
-    let mut merged_boundaries = loop_state.boundaries;
-
-    let (
-        start,
-        end,
-        start_sources,
-        end_sources,
-        start_bounds,
-        end_bounds,
-        inclusive,
-        step,
-        step_op,
-        reverse,
-    ) = match &for_stmt.range {
-        ForRange::Forward {
-            start: range_start,
-            end: range_end,
-            inclusive,
-            step,
-        } => {
-            let (start, start_sources, start_bounds) =
-                eval_for_bound(module, &store, range_start, arena)?;
-            let (end, end_sources, end_bounds) = eval_for_bound(module, &store, range_end, arena)?;
-            (
-                start,
-                end,
-                start_sources,
-                end_sources,
-                start_bounds,
-                end_bounds,
-                *inclusive,
-                *step,
-                SLTStepOp::Add,
-                false,
-            )
-        }
-        ForRange::Reverse {
-            start: range_start,
-            end: range_end,
-            inclusive,
-            step,
-        } => {
-            let (start, start_sources, start_bounds) =
-                eval_for_bound(module, &store, range_start, arena)?;
-            let (end, end_sources, end_bounds) = eval_for_bound(module, &store, range_end, arena)?;
-            (
-                start,
-                end,
-                start_sources,
-                end_sources,
-                start_bounds,
-                end_bounds,
-                *inclusive,
-                *step,
-                SLTStepOp::Add,
-                true,
-            )
-        }
-        ForRange::Stepped {
-            start: range_start,
-            end: range_end,
-            inclusive,
-            step,
-            op,
-        } => {
-            let (start, start_sources, start_bounds) =
-                eval_for_bound(module, &store, range_start, arena)?;
-            let (end, end_sources, end_bounds) = eval_for_bound(module, &store, range_end, arena)?;
-            let step_op = match op {
-                Op::Mul => SLTStepOp::Mul,
-                Op::LogicShiftL | Op::ArithShiftL => SLTStepOp::Shl,
-                other => {
-                    return Err(ParserError::unsupported(
-                        65,
-                        LoweringPhase::CombLowering,
-                        "for loop step operator",
-                        format!("{other:?}"),
-                        Some(&for_stmt.token),
-                    ));
-                }
-            };
-            (
-                start,
-                end,
-                start_sources,
-                end_sources,
-                start_bounds,
-                end_bounds,
-                *inclusive,
-                *step,
-                step_op,
-                false,
-            )
-        }
-    };
-
-    merged_boundaries = merge_boundaries(merged_boundaries, start_bounds);
-    merged_boundaries = merge_boundaries(merged_boundaries, end_bounds);
-
-    let updates = extract_store_updates(&iter_store_before, &iter_store_after, arena)?;
-    if updates.is_empty() && effects.is_empty() {
-        let mut store = store;
-        store.remove(&for_stmt.var_id);
-        return Ok((store, merged_boundaries, None));
-    }
-
-    let folded_updates: Vec<_> = if updates.is_empty() {
-        let one = bool_node(arena, true)?;
-        vec![SLTForUpdate {
-            target: VarAtomBase::new(for_stmt.var_id, 0, loop_width - 1),
-            expr: one,
-        }]
-    } else {
-        updates
-            .iter()
-            .map(|(target, expr, _)| SLTForUpdate {
-                target: *target,
-                expr: *expr,
-            })
-            .collect()
-    };
-    let loop_updated_vars: HashSet<_> = folded_updates
-        .iter()
-        .map(|update| update.target.id)
-        .collect();
-    let (initial_updates, initial_sources): (Vec<_>, HashSet<_>) = if updates.is_empty() {
-        let one = bool_node(arena, true)?;
-        (
-            vec![SLTForUpdate {
-                target: VarAtomBase::new(for_stmt.var_id, 0, loop_width - 1),
-                expr: one,
-            }],
-            HashSet::default(),
-        )
-    } else {
-        let mut initial_sources = HashSet::default();
-        let initial_updates = updates
-            .iter()
-            .map(|(target, _, _)| {
-                let range_store = store.get(&target.id).ok_or_else(|| {
-                    ParserError::illegal_context(
-                        "for-loop initial state",
-                        "state variable is absent from the symbolic store",
-                        Some(&for_stmt.token),
-                    )
-                })?;
-                let parts = range_store.get_parts(target.access).map_err(|error| {
-                    range_store_error("for-loop initial state", error, Some(&for_stmt.token))
-                })?;
-                let (expr, sources) =
-                    combine_parts_with_default(target.id, target.access.lsb, parts, arena)?;
-                initial_sources.extend(sources);
-                Ok(SLTForUpdate {
-                    target: *target,
-                    expr,
-                })
-            })
-            .collect::<Result<Vec<_>, ParserError>>()?;
-        (initial_updates, initial_sources)
-    };
-
-    let loop_runner = if effects.is_empty() {
-        None
-    } else {
-        let result = folded_updates[0].target;
-        Some(arena.alloc(SLTNode::ForFold {
-            loop_var: for_stmt.var_id,
-            loop_width,
-            loop_signed: for_stmt.var_type.signed,
-            start: start.clone(),
-            end: end.clone(),
-            inclusive,
-            step,
-            step_op,
-            reverse,
-            result,
-            initials: initial_updates.clone(),
-            updates: folded_updates.clone(),
-            effects: effects.to_vec(),
-            continue_cond: loop_state.continue_expr,
-        })?)
-    };
-
-    if updates.is_empty() {
-        let mut store = store;
-        store.remove(&for_stmt.var_id);
-        return Ok((store, merged_boundaries, loop_runner));
-    }
-
-    let mut result_store = store;
-    for (target, _expr, sources) in updates {
-        let mut all_sources = start_sources.clone();
-        all_sources.extend(end_sources.iter().copied());
-        all_sources.extend(
-            loop_state
-                .continue_sources
-                .iter()
-                .copied()
-                .filter(|src| src.id != for_stmt.var_id && !loop_updated_vars.contains(&src.id)),
-        );
-        all_sources.extend(
-            sources
-                .into_iter()
-                .filter(|src| src.id != for_stmt.var_id && !loop_updated_vars.contains(&src.id)),
-        );
-        // The fold body reads loop-carried values, but their initial values
-        // may come from external state which is absent from every body update
-        // source after carried-state filtering.  Keep those dependencies so
-        // hierarchy glue and preceding procedural writes are scheduled before
-        // the fold.
-        all_sources.extend(
-            initial_sources
-                .iter()
-                .copied()
-                .filter(|src| src.id != for_stmt.var_id),
-        );
-        all_sources.retain(|src| src.id != target.id);
-
-        let folded_expr = arena.alloc(SLTNode::ForFold {
-            loop_var: for_stmt.var_id,
-            loop_width,
-            loop_signed: for_stmt.var_type.signed,
-            start: start.clone(),
-            end: end.clone(),
-            inclusive,
-            step,
-            step_op,
-            reverse,
-            result: target,
-            initials: initial_updates.clone(),
-            updates: folded_updates.clone(),
-            effects: Vec::new(),
-            continue_cond: loop_state.continue_expr,
-        })?;
-
-        let variable = module.variables.get(&target.id).ok_or_else(|| {
-            ParserError::illegal_context(
-                "for-loop result state",
-                "state variable is absent from the semantic module",
-                Some(&for_stmt.token),
-            )
-        })?;
-        let width = resolve_total_width(module, variable)?;
-        result_store
-            .entry(target.id)
-            .or_insert_with(|| RangeStore::new(None, width))
-            .update(target.access, Some((folded_expr, all_sources)))
-            .map_err(|error| {
-                range_store_error("for-loop result state", error, Some(&for_stmt.token))
-            })?;
-    }
-
-    result_store.remove(&for_stmt.var_id);
-    Ok((result_store, merged_boundaries, loop_runner))
-}
-
-fn checked_destination_width(
-    module: &Module,
-    destinations: &[veryl_analyzer::ir::AssignDestination],
-    context: &'static str,
-    token: Option<&TokenRange>,
-) -> Result<usize, ParserError> {
-    if destinations.is_empty() {
-        return Err(ParserError::illegal_context(
-            context,
-            "assignment has no destination",
-            token,
-        ));
-    }
-    let mut total = 0usize;
-    for destination in destinations {
-        let width = crate::parser::bitaccess::get_access_width(
-            module,
-            destination.id,
-            &destination.index,
-            &destination.select,
-        )?;
-        total = total.checked_add(width).ok_or_else(|| {
-            ParserError::illegal_context(
-                context,
-                "concatenated destination width overflows usize",
-                Some(&destination.token),
-            )
-        })?;
-    }
-    if total == 0 {
-        return Err(ParserError::illegal_context(
-            context,
-            "assignment destination has zero width",
-            token,
-        ));
-    }
-    Ok(total)
-}
-
-fn checked_assignment_slice(
-    offset: usize,
-    width: usize,
-    rhs_width: usize,
-    destination: &veryl_analyzer::ir::AssignDestination,
-) -> Result<(BitAccess, usize), ParserError> {
-    let end = offset.checked_add(width).ok_or_else(|| {
-        ParserError::illegal_context(
-            "concatenated assignment",
-            "RHS slice end overflows usize",
-            Some(&destination.token),
-        )
-    })?;
-    if width == 0 || end > rhs_width {
-        return Err(ParserError::illegal_context(
-            "concatenated assignment",
-            format!("RHS slice {offset}..{end} is outside width {rhs_width}"),
-            Some(&destination.token),
-        ));
-    }
-    Ok((BitAccess::new(offset, end - 1), end))
-}
-
-fn record_assignment_boundary(
-    boundaries: &mut BoundaryMap<VarId>,
-    destination: &veryl_analyzer::ir::AssignDestination,
-    access: BitAccess,
-) -> Result<(), ParserError> {
-    let end = access.msb.checked_add(1).ok_or_else(|| {
-        ParserError::illegal_context(
-            "assignment destination",
-            "destination boundary overflows usize",
-            Some(&destination.token),
-        )
-    })?;
-    let entry = boundaries.entry(destination.id).or_default();
-    entry.insert(access.lsb);
-    entry.insert(end);
-    Ok(())
-}
-
-fn update_assignment_range(
-    module: &Module,
-    store: &mut SymbolicStore<VarId>,
-    destination: &veryl_analyzer::ir::AssignDestination,
-    access: BitAccess,
-    mut value: (NodeId, HashSet<VarAtomBase<VarId>>),
-    source_is_2state: bool,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(), ParserError> {
-    let variable = module.variables.get(&destination.id).ok_or_else(|| {
-        ParserError::illegal_context(
-            "assignment destination",
-            "destination variable is absent from the semantic module",
-            Some(&destination.token),
-        )
-    })?;
-    let variable_width = resolve_total_width(module, variable)?;
-    if variable_width == 0 || access.msb >= variable_width {
-        return Err(ParserError::illegal_context(
-            "assignment destination",
-            format!(
-                "destination access [{}:{}] is outside variable width {variable_width}",
-                access.msb, access.lsb
-            ),
-            Some(&destination.token),
-        ));
-    }
-    if variable.r#type.is_2state() && !source_is_2state {
-        value.0 = arena.alloc(SLTNode::Unary(UnaryOp::ToTwoState, value.0))?;
-    }
-    let range_store = store.get_mut(&destination.id).ok_or_else(|| {
-        ParserError::illegal_context(
-            "assignment destination",
-            "destination variable is absent from the symbolic store",
-            Some(&destination.token),
-        )
-    })?;
-    range_store.update(access, Some(value)).map_err(|error| {
-        range_store_error("assignment destination", error, Some(&destination.token))
-    })?;
-    Ok(())
-}
-
-fn eval_assign(
-    module: &Module,
-    mut store: SymbolicStore<VarId>,
-    boundaries: BoundaryMap<VarId>,
-    stmt: &AssignStatement,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-    let rhs_is_2state = stmt.expr.comptime().r#type.is_2state();
-    let rhs_expected_width = checked_destination_width(
-        module,
-        &stmt.dst,
-        "assignment destination",
-        Some(&stmt.expr.token_range()),
-    )?;
-    let ((rhs_expr, rhs_sources), rhs_bounds) = if let Expression::ArrayLiteral(items, _) =
-        &stmt.expr
-    {
-        let ((node, sources), bounds) =
-            eval_array_literal_expression(module, &store, items, Some(rhs_expected_width), arena)?;
-        if get_width(node, arena) == 0 {
-            return Err(ParserError::illegal_context(
-                "assignment expression",
-                "a zero-width array literal cannot be assigned",
-                Some(&stmt.expr.token_range()),
-            ));
-        }
-        (
-            (
-                coerce_node_width(arena, node, Some(rhs_expected_width), false)?,
-                sources,
-            ),
-            bounds,
-        )
-    } else {
-        eval_assignment_expression(module, &store, &stmt.expr, arena, rhs_expected_width)?
-    };
-    let mut boundaries = merge_boundaries(boundaries, rhs_bounds);
-
-    if stmt.dst.len() == 1 {
-        // Single destination: store RHS directly
-        let dst = &stmt.dst[0];
-
-        if crate::parser::bitaccess::is_static_access(&dst.index, &dst.select) {
-            let access = eval_var_select(module, dst.id, &dst.index, &dst.select)?;
-
-            record_assignment_boundary(&mut boundaries, dst, access)?;
-            update_assignment_range(
-                module,
-                &mut store,
-                dst,
-                access,
-                (rhs_expr, rhs_sources.clone()),
-                rhs_is_2state,
-                arena,
-            )?;
-        } else {
-            let (s, b) = eval_dynamic_assign(
-                module,
-                store,
-                boundaries,
-                dst,
-                rhs_expr,
-                rhs_sources.clone(),
-                rhs_is_2state,
-                arena,
-            )?;
-            return Ok((s, b));
-        }
-    } else {
-        // LHS concatenation: slice RHS for each destination
-        // dst is ordered MSB-first (e.g., {a, b} means a=MSB, b=LSB),
-        // so iterate in reverse to compute offsets from LSB.
-        let mut current_offset = 0;
-        for dst in stmt.dst.iter().rev() {
-            let part_width = crate::parser::bitaccess::get_access_width(
-                module,
-                dst.id,
-                &dst.index,
-                &dst.select,
-            )?;
-
-            // Slice the RHS to extract the bits for this destination
-            let (slice_access, next_offset) =
-                checked_assignment_slice(current_offset, part_width, rhs_expected_width, dst)?;
-            let slice_expr = arena.alloc(SLTNode::Slice {
-                expr: rhs_expr,
-                access: slice_access,
-            })?;
-
-            if crate::parser::bitaccess::is_static_access(&dst.index, &dst.select) {
-                let access = eval_var_select(module, dst.id, &dst.index, &dst.select)?;
-
-                record_assignment_boundary(&mut boundaries, dst, access)?;
-                update_assignment_range(
-                    module,
-                    &mut store,
-                    dst,
-                    access,
-                    (slice_expr, rhs_sources.clone()),
-                    rhs_is_2state,
-                    arena,
-                )?;
-            } else {
-                let (s, b) = eval_dynamic_assign(
-                    module,
-                    store,
-                    boundaries,
-                    dst,
-                    slice_expr,
-                    rhs_sources.clone(),
-                    rhs_is_2state,
-                    arena,
-                )?;
-                store = s;
-                boundaries = b;
-            }
-
-            current_offset = next_offset;
-        }
-        if current_offset != rhs_expected_width {
-            return Err(ParserError::illegal_context(
-                "concatenated assignment",
-                format!(
-                    "destinations cover {current_offset} bits, but the RHS has width {rhs_expected_width}"
-                ),
-                Some(&stmt.expr.token_range()),
-            ));
-        }
-    }
-    Ok((store, boundaries))
-}
-
-fn assign_node_to_dsts(
-    module: &Module,
-    mut store: SymbolicStore<VarId>,
-    mut boundaries: BoundaryMap<VarId>,
-    dsts: &[veryl_analyzer::ir::AssignDestination],
-    rhs_expr: NodeId,
-    rhs_sources: HashSet<VarAtomBase<VarId>>,
-    source_is_2state: bool,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-    let destination_width = checked_destination_width(
-        module,
-        dsts,
-        "function output destination",
-        dsts.first().map(|destination| &destination.token),
-    )?;
-    let rhs_width = get_width(rhs_expr, arena);
-    if rhs_width == 0 {
-        return Err(ParserError::illegal_context(
-            "function output destination",
-            "function output value has zero width",
-            dsts.first().map(|destination| &destination.token),
-        ));
-    }
-    let rhs_signed = expr::is_signed(module, rhs_expr, arena);
-    let rhs_expr = coerce_node_width(arena, rhs_expr, Some(destination_width), rhs_signed)?;
-
-    if dsts.len() == 1 {
-        let dst = &dsts[0];
-        if crate::parser::bitaccess::is_static_access(&dst.index, &dst.select) {
-            let access = eval_var_select(module, dst.id, &dst.index, &dst.select)?;
-            record_assignment_boundary(&mut boundaries, dst, access)?;
-            update_assignment_range(
-                module,
-                &mut store,
-                dst,
-                access,
-                (rhs_expr, rhs_sources),
-                source_is_2state,
-                arena,
-            )?;
-
-            return Ok((store, boundaries));
-        }
-
-        return eval_dynamic_assign(
-            module,
-            store,
-            boundaries,
-            dst,
-            rhs_expr,
-            rhs_sources,
-            source_is_2state,
-            arena,
-        );
-    }
-
-    let mut current_offset = 0;
-    for dst in dsts.iter().rev() {
-        let part_width =
-            crate::parser::bitaccess::get_access_width(module, dst.id, &dst.index, &dst.select)?;
-        let (slice_access, next_offset) =
-            checked_assignment_slice(current_offset, part_width, destination_width, dst)?;
-        let slice_expr = arena.alloc(SLTNode::Slice {
-            expr: rhs_expr,
-            access: slice_access,
-        })?;
-
-        if crate::parser::bitaccess::is_static_access(&dst.index, &dst.select) {
-            let access = eval_var_select(module, dst.id, &dst.index, &dst.select)?;
-
-            record_assignment_boundary(&mut boundaries, dst, access)?;
-            update_assignment_range(
-                module,
-                &mut store,
-                dst,
-                access,
-                (slice_expr, rhs_sources.clone()),
-                source_is_2state,
-                arena,
-            )?;
-        } else {
-            let (next_store, next_boundaries) = eval_dynamic_assign(
-                module,
-                store,
-                boundaries,
-                dst,
-                slice_expr,
-                rhs_sources.clone(),
-                source_is_2state,
-                arena,
-            )?;
-            store = next_store;
-            boundaries = next_boundaries;
-        }
-
-        current_offset = next_offset;
-    }
-
-    if current_offset != destination_width {
-        return Err(ParserError::illegal_context(
-            "function output destination",
-            format!(
-                "destinations cover {current_offset} bits, but the output value has width {destination_width}"
-            ),
-            dsts.first().map(|destination| &destination.token),
-        ));
-    }
-
-    Ok((store, boundaries))
-}
-
-fn eval_statement_form_function_call(
-    module: &Module,
-    mut store: SymbolicStore<VarId>,
-    mut boundaries: BoundaryMap<VarId>,
-    call: &veryl_analyzer::ir::FunctionCall,
-    arena: &mut SLTNodeArena<VarId>,
-    phase: LoweringPhase,
-) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-    if call.outputs.is_empty() {
-        return Err(ParserError::unsupported(
-            58,
-            phase,
-            "statement-form function call without output arguments",
-            format!("{call}"),
-            Some(&call.comptime.token),
-        ));
-    }
-
-    let Some(function) = module.functions.get(&call.id) else {
-        return Err(ParserError::unsupported(
-            60,
-            phase,
-            "function call",
-            format!("unknown function id: {:?}", call.id),
-            Some(&call.comptime.token),
-        ));
-    };
-
-    let Some(function_body) = (if let Some(index) = &call.index {
-        function.get_function(index)
-    } else {
-        function.get_function(&[])
-    }) else {
-        return Err(ParserError::unsupported(
-            60,
-            phase,
-            "function call specialization",
-            format!("{call}"),
-            Some(&call.comptime.token),
-        ));
-    };
-
-    let mut local_store = store.clone();
-
-    for (arg_path, arg_id) in &function_body.arg_map {
-        let Some(arg_expr) = call.inputs.get(arg_path) else {
-            if call.outputs.contains_key(arg_path) {
-                continue;
-            }
-            return Err(ParserError::unsupported(
-                61,
-                phase,
-                "function call missing argument",
-                format!("{call}"),
-                Some(&call.comptime.token),
-            ));
-        };
-
-        let formal = module.variables.get(arg_id).ok_or_else(|| {
-            ParserError::illegal_context(
-                "function input argument",
-                "formal variable is absent from the semantic module",
-                Some(&call.comptime.token),
-            )
-        })?;
-        let arg_width = resolve_total_width(module, formal)?;
-        let ((arg_node, arg_sources), arg_bounds) =
-            eval_assignment_expression(module, &store, arg_expr, arena, arg_width)?;
-        let arg_node = if formal.r#type.is_2state() && !arg_expr.comptime().r#type.is_2state() {
-            arena.alloc(SLTNode::Unary(UnaryOp::ToTwoState, arg_node))?
-        } else {
-            arg_node
-        };
-        boundaries = merge_boundaries(boundaries, arg_bounds);
-        local_store.insert(
-            *arg_id,
-            RangeStore::new(Some((arg_node, arg_sources)), arg_width),
-        );
-    }
-
-    let (final_local_store, local_boundaries) = if let Some(ret_id) = function_body.ret {
-        let ((_, _), local_boundaries, final_local_store) =
-            eval_function_body_return(module, &local_store, &function_body, ret_id, arena)?;
-        (final_local_store, local_boundaries)
-    } else {
-        function_body.statements.iter().try_fold(
-            (local_store, BoundaryMap::default()),
-            |(local_store, local_boundaries), stmt| {
-                eval_statement(module, local_store, local_boundaries, stmt, arena)
-            },
-        )?
-    };
-    boundaries = merge_boundaries(boundaries, local_boundaries);
-
-    for (arg_path, dsts) in &call.outputs {
-        let Some(arg_id) = function_body.arg_map.get(arg_path) else {
-            return Err(ParserError::unsupported(
-                61,
-                phase,
-                "function call missing argument",
-                format!("{call}"),
-                Some(&call.comptime.token),
-            ));
-        };
-
-        let formal = module.variables.get(arg_id).ok_or_else(|| {
-            ParserError::illegal_context(
-                "function output value",
-                "formal variable is absent from the semantic module",
-                Some(&call.comptime.token),
-            )
-        })?;
-        let formal_width = resolve_total_width(module, formal)?;
-        if formal_width == 0 {
-            return Err(ParserError::illegal_context(
-                "function output value",
-                "formal output has zero width",
-                Some(&call.comptime.token),
-            ));
-        }
-        let access = BitAccess::new(0, formal_width - 1);
-        let range_store = final_local_store.get(arg_id).ok_or_else(|| {
-            ParserError::illegal_context(
-                "function output value",
-                "formal output is absent from the final symbolic store",
-                Some(&call.comptime.token),
-            )
-        })?;
-        let parts = range_store.get_parts(access).map_err(|error| {
-            range_store_error("function output value", error, Some(&call.comptime.token))
-        })?;
-        let (output_expr, output_sources) = combine_parts_with_default(*arg_id, 0, parts, arena)?;
-        let (next_store, next_boundaries) = assign_node_to_dsts(
-            module,
-            store,
-            boundaries,
-            dsts,
-            output_expr,
-            output_sources,
-            formal.r#type.is_2state(),
-            arena,
-        )?;
-        store = next_store;
-        boundaries = next_boundaries;
-    }
-
-    Ok((store, boundaries))
-}
-
-struct DynamicSelectOffset {
-    node: NodeId,
-    indices: Vec<SLTIndex>,
-    sources: HashSet<VarAtomBase<VarId>>,
-    boundaries: BoundaryMap<VarId>,
-}
-
-/// Build the effective LSB for a dynamic access from validated select
-/// geometry.  The returned `indices` and arithmetic `node` encode the same
-/// offset so direct dynamic loads and read-modify-write paths cannot diverge.
-fn eval_dynamic_select_offset(
-    module: &Module,
-    store: &SymbolicStore<VarId>,
-    var_id: VarId,
-    index: &VarIndex,
-    select: &VarSelect,
-    arena: &mut SLTNodeArena<VarId>,
-    token: Option<&TokenRange>,
-) -> Result<DynamicSelectOffset, ParserError> {
-    let geometry = select_geometry(module, var_id, index, select)?;
-    let array_dimension_count = module.variables[&var_id].r#type.array.iter().count();
-    let array_element_width = if array_dimension_count == 0 {
-        None
-    } else {
-        geometry.strides.get(array_dimension_count - 1).copied()
-    };
-    let mut offset = arena.alloc(SLTNode::Constant(
-        BigUint::from(0u8),
-        BigUint::from(0u8),
-        64,
-        false,
-    ))?;
-    let mut indices = Vec::new();
-    let mut sources = HashSet::default();
-    let mut boundaries = BoundaryMap::default();
-
-    let mut expressions = index.0.clone();
-    expressions.extend(select.0.clone());
-    for (dimension, expression) in expressions[..geometry.dimension_count].iter().enumerate() {
-        let ((node, node_sources), node_boundaries) =
-            eval_expression(module, store, expression, arena, None)?;
-        sources.extend(node_sources);
-        boundaries = merge_boundaries(boundaries, node_boundaries);
-        let stride = geometry.strides.get(dimension).copied().ok_or_else(|| {
-            ParserError::illegal_context(
-                "dynamic variable select",
-                format!(
-                    "index dimension {dimension} is outside the {}-entry stride table",
-                    geometry.strides.len()
-                ),
-                token,
-            )
-        })?;
-        let kind = if dimension < array_dimension_count {
-            SLTIndexKind::Unpacked {
-                element_width: array_element_width.expect("unpacked array has an element width"),
-            }
-        } else {
-            SLTIndexKind::Packed
-        };
-        indices.push(SLTIndex { node, stride, kind });
-        let stride_node = arena.alloc(SLTNode::Constant(
-            BigUint::from(stride),
-            BigUint::from(0u8),
-            64,
-            false,
-        ))?;
-        let term = arena.alloc(SLTNode::Binary(node, BinaryOp::Mul, stride_node))?;
-        offset = arena.alloc(SLTNode::Binary(offset, BinaryOp::Add, term))?;
-    }
-
-    if let Some(part) = geometry.part {
-        let stride = geometry
-            .strides
-            .get(geometry.dimension_count)
-            .copied()
-            .ok_or_else(|| {
-                ParserError::illegal_context(
-                    "dynamic variable select",
-                    format!(
-                        "part-select dimension {} is outside the {}-entry stride table",
-                        geometry.dimension_count,
-                        geometry.strides.len()
-                    ),
-                    token,
-                )
-            })?;
-        let start = match part {
-            PartSelectGeometry::Colon { lsb, .. } => arena.alloc(SLTNode::Constant(
-                BigUint::from(lsb),
-                BigUint::from(0u8),
-                64,
-                false,
-            ))?,
-            PartSelectGeometry::PlusColon { .. }
-            | PartSelectGeometry::MinusColon { .. }
-            | PartSelectGeometry::Step { .. } => {
-                let anchor_expression = select.0.last().ok_or_else(|| {
-                    ParserError::illegal_context(
-                        "dynamic variable select",
-                        "part select is missing its anchor expression",
-                        token,
-                    )
-                })?;
-                let ((anchor, anchor_sources), anchor_boundaries) =
-                    eval_expression(module, store, anchor_expression, arena, None)?;
-                sources.extend(anchor_sources);
-                boundaries = merge_boundaries(boundaries, anchor_boundaries);
-                match part {
-                    PartSelectGeometry::PlusColon { .. } => anchor,
-                    PartSelectGeometry::MinusColon { elements } => {
-                        let decrement = elements.checked_sub(1).ok_or_else(|| {
-                            ParserError::illegal_context(
-                                "dynamic variable select",
-                                "minus-colon width underflows",
-                                token,
-                            )
-                        })?;
-                        let decrement = arena.alloc(SLTNode::Constant(
-                            BigUint::from(decrement),
-                            BigUint::from(0u8),
-                            64,
-                            false,
-                        ))?;
-                        arena.alloc(SLTNode::Binary(anchor, BinaryOp::Sub, decrement))?
-                    }
-                    PartSelectGeometry::Step { elements } => {
-                        let elements = arena.alloc(SLTNode::Constant(
-                            BigUint::from(elements),
-                            BigUint::from(0u8),
-                            64,
-                            false,
-                        ))?;
-                        arena.alloc(SLTNode::Binary(anchor, BinaryOp::Mul, elements))?
-                    }
-                    PartSelectGeometry::Colon { .. } => {
-                        return Err(ParserError::illegal_context(
-                            "dynamic variable select",
-                            "inconsistent colon-select geometry",
-                            token,
-                        ));
-                    }
-                }
-            }
-        };
-        indices.push(SLTIndex {
-            node: start,
-            stride,
-            kind: SLTIndexKind::Packed,
-        });
-        let stride_node = arena.alloc(SLTNode::Constant(
-            BigUint::from(stride),
-            BigUint::from(0u8),
-            64,
-            false,
-        ))?;
-        let term = arena.alloc(SLTNode::Binary(start, BinaryOp::Mul, stride_node))?;
-        offset = arena.alloc(SLTNode::Binary(offset, BinaryOp::Add, term))?;
-    }
-
-    Ok(DynamicSelectOffset {
-        node: offset,
-        indices,
-        sources,
-        boundaries,
-    })
-}
-
-fn eval_dynamic_assign(
-    module: &Module,
-    mut store: SymbolicStore<VarId>,
-    mut boundaries: BoundaryMap<VarId>,
-    dst: &veryl_analyzer::ir::AssignDestination,
-    rhs_expr: NodeId,
-    rhs_sources: HashSet<VarAtomBase<VarId>>,
-    source_is_2state: bool,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(SymbolicStore<VarId>, BoundaryMap<VarId>), ParserError> {
-    let mut all_sources = rhs_sources;
-    let select_offset = eval_dynamic_select_offset(
-        module,
-        &store,
-        dst.id,
-        &dst.index,
-        &dst.select,
-        arena,
-        Some(&dst.token),
-    )?;
-    boundaries = merge_boundaries(boundaries, select_offset.boundaries);
-    all_sources.extend(select_offset.sources);
-    let offset_node = select_offset.node;
-
-    let access_width =
-        crate::parser::bitaccess::get_access_width(module, dst.id, &dst.index, &dst.select)?;
-    let var = &module.variables[&dst.id];
-    let width = resolve_total_width(module, var)?;
-    if width == 0 || access_width == 0 || access_width > width {
-        return Err(ParserError::illegal_context(
-            "dynamic assignment",
-            format!("destination width {access_width} must be in 1..={width}"),
-            Some(&dst.token),
-        ));
-    }
-
-    let access_full = BitAccess::new(0, width - 1);
-    let range_store = store
-        .entry(dst.id)
-        .or_insert_with(|| RangeStore::new(None, width));
-
-    // Evaluate the variable's current state.
-    // Sub-ranges that haven't been assigned yet will fall back to their initial input state.
-    let old_parts = range_store
-        .get_parts(access_full)
-        .map_err(|error| range_store_error("dynamic assignment", error, Some(&dst.token)))?;
-    let (old_val, old_sources) = combine_parts_with_default(dst.id, 0, old_parts, arena)?;
-    // Note: Partial dynamic updates are not treated as self-dependencies (latches)
-    // to maintain consistency with existing test expectations and Verilog semantics.
-    for source in old_sources {
-        if source.id != dst.id {
-            all_sources.insert(source);
-        }
-    }
-
-    // Compute the bitmask to isolate the target range: mask = !(( (1<<access_width) - 1 ) << offset)
-    let mask_base = (BigUint::from(1u32) << access_width) - BigUint::from(1u32);
-    // Ensure width consistency; using the full variable width for safety.
-    let mask_constant = arena.alloc(SLTNode::Constant(
-        mask_base,
-        BigUint::from(0u32),
-        width,
-        false,
-    ))?;
-
-    let mask_shifted = arena.alloc(SLTNode::Binary(mask_constant, BinaryOp::Shl, offset_node))?;
-    let mask_node = arena.alloc(SLTNode::Unary(UnaryOp::BitNot, mask_shifted))?;
-
-    // Apply assignment coercion before embedding the value in the full
-    // destination.  Otherwise discarded high RHS bits can corrupt neighbours.
-    let rhs_signed = expr::is_signed(module, rhs_expr, arena);
-    let rhs_expr = coerce_node_width(arena, rhs_expr, Some(access_width), rhs_signed)?;
-    let rhs_expr = if var.r#type.is_2state() && !source_is_2state {
-        arena.alloc(SLTNode::Unary(UnaryOp::ToTwoState, rhs_expr))?
-    } else {
-        rhs_expr
-    };
-    let rhs_widened = if access_width < width {
-        let padding = width - access_width;
-        let zero = arena.alloc(SLTNode::Constant(
-            BigUint::from(0u32),
-            BigUint::from(0u32),
-            padding,
-            false,
-        ))?;
-        // Concatenate zero padding to match variable width: {padding'b0, rhs_expr}
-        arena.alloc(SLTNode::Concat(vec![
-            (zero, padding),
-            (rhs_expr, access_width),
-        ]))?
-    } else {
-        rhs_expr
-    };
-    let new_val_term = arena.alloc(SLTNode::Binary(rhs_widened, BinaryOp::Shl, offset_node))?;
-    let new_val_term = arena.alloc(SLTNode::Binary(new_val_term, BinaryOp::And, mask_shifted))?;
-
-    // Apply the update: final_val = (old_val & mask) | new_val_term
-    let new_val_masked = arena.alloc(SLTNode::Binary(old_val, BinaryOp::And, mask_node))?;
-    let final_val = arena.alloc(SLTNode::Binary(new_val_masked, BinaryOp::Or, new_val_term))?;
-
-    let prefix_access = eval_var_select(module, dst.id, &dst.index, &dst.select)?;
-    let stored_expr = if prefix_access.lsb == 0 && prefix_access.msb == width - 1 {
-        final_val
-    } else {
-        arena.alloc(SLTNode::Slice {
-            expr: final_val,
-            access: prefix_access,
-        })?
-    };
-    range_store
-        .update(prefix_access, Some((stored_expr, all_sources)))
-        .map_err(|error| range_store_error("dynamic assignment", error, Some(&dst.token)))?;
-
-    Ok((store, boundaries))
-}
-fn eval_if_with_recovery(
-    module: &Module,
-    initial_store: SymbolicStore<VarId>,
-    mut boundaries: HashMap<VarId, BTreeSet<usize>>,
-    stmt: &IfStatement,
-    arena: &mut SLTNodeArena<VarId>,
-    loop_candidates: &[LoopRecoveryCandidate],
-    active_guard: Option<&ActiveGuard>,
-) -> Result<(SymbolicStore<VarId>, HashMap<VarId, BTreeSet<usize>>), ParserError> {
-    let ((cond_expr, cond_sources), cond_bounds) =
-        eval_expression(module, &initial_store, &stmt.cond, arena, None)?;
-    let cond_expr = procedural_condition(arena, cond_expr)?;
-    boundaries.extend(cond_bounds);
-
-    // Constant folding: if condition is a constant, inline the appropriate side
-    if let SLTNode::Constant(val, _, _, _) = arena.get(cond_expr) {
-        let side = if *val != BigUint::from(0u32) {
-            &stmt.true_side
-        } else {
-            &stmt.false_side
-        };
-        return eval_statements_with_recovery(
-            module,
-            initial_store,
-            boundaries,
-            side,
-            arena,
-            loop_candidates,
-            active_guard,
-        );
-    }
-
-    // Evaluate Then and Else paths independently
-    let then_guard = combine_active_guard(arena, active_guard, cond_expr, &cond_sources)?;
-    let false_condition = invert_active_condition(arena, cond_expr)?;
-    let else_guard = combine_active_guard(arena, active_guard, false_condition, &cond_sources)?;
-    let (then_store, b_then) = eval_statements_with_recovery(
-        module,
-        initial_store.clone(),
-        boundaries.clone(),
-        &stmt.true_side,
-        arena,
-        loop_candidates,
-        Some(&then_guard),
-    )?;
-    let (else_store, b_else) = eval_statements_with_recovery(
-        module,
-        initial_store,
-        b_then,
-        &stmt.false_side,
-        arena,
-        loop_candidates,
-        Some(&else_guard),
-    )?;
-
-    Ok((
-        merge_symbolic_stores(
-            module,
-            &then_store,
-            &else_store,
-            cond_expr,
-            &cond_sources,
-            arena,
-        )?,
-        b_else,
-    ))
-}
-
-fn eval_if(
-    module: &Module,
-    initial_store: SymbolicStore<VarId>,
-    mut boundaries: HashMap<VarId, BTreeSet<usize>>,
-    stmt: &IfStatement,
-    arena: &mut SLTNodeArena<VarId>,
-) -> Result<(SymbolicStore<VarId>, HashMap<VarId, BTreeSet<usize>>), ParserError> {
-    let ((cond_expr, cond_sources), cond_bounds) =
-        eval_expression(module, &initial_store, &stmt.cond, arena, None)?;
-    let cond_expr = procedural_condition(arena, cond_expr)?;
-    boundaries.extend(cond_bounds);
-
-    if let SLTNode::Constant(val, _, _, _) = arena.get(cond_expr) {
-        let side = if *val != BigUint::from(0u32) {
-            &stmt.true_side
-        } else {
-            &stmt.false_side
-        };
-        return side
-            .iter()
-            .try_fold((initial_store, boundaries), |(s, b), step| {
-                eval_statement(module, s, b, step, arena)
-            });
-    }
-
-    let (then_store, b_then) = stmt.true_side.iter().try_fold(
-        (initial_store.clone(), boundaries.clone()),
-        |(s, b), step| eval_statement(module, s, b, step, arena),
-    )?;
-    let (else_store, b_else) = stmt
-        .false_side
-        .iter()
-        .try_fold((initial_store, b_then), |(s, b), step| {
-            eval_statement(module, s, b, step, arena)
-        })?;
-
-    Ok((
-        merge_symbolic_stores(
-            module,
-            &then_store,
-            &else_store,
-            cond_expr,
-            &cond_sources,
-            arena,
-        )?,
-        b_else,
-    ))
-}
-
-fn combine_parts_with_default<A: Clone + PartialEq + Eq + Hash>(
-    var_id: A,
-    start_lsb: usize,
-    parts: Vec<(Option<(NodeId, HashSet<VarAtomBase<A>>)>, BitAccess)>,
-    arena: &mut SLTNodeArena<A>,
-) -> Result<(NodeId, HashSet<VarAtomBase<A>>), SLTNodeFactsError> {
-    let mut fixed_parts = Vec::new();
-    let mut current_lsb = start_lsb;
-    for (val_opt, access) in parts {
-        let width = access.msb - access.lsb + 1;
-        match val_opt {
-            Some((expr, s)) => {
-                fixed_parts.push(((expr, s), access));
-            }
-            None => {
-                let input_node = arena.alloc(SLTNode::Input {
-                    variable: var_id.clone(),
-                    signed: false,
-                    index: vec![],
-                    access: BitAccess::new(current_lsb, current_lsb + width - 1),
-                })?;
-                let mut sources = HashSet::default();
-                sources.insert(VarAtomBase::new(
-                    var_id.clone(),
-                    current_lsb,
-                    current_lsb + width - 1,
-                ));
-                fixed_parts.push(((input_node, sources), BitAccess::new(0, width - 1)));
-            }
-        }
-        current_lsb += width;
-    }
-    combine_parts(fixed_parts, arena)
-}
-
-fn combine_parts<A: Clone + PartialEq + Eq + Hash>(
-    parts: Vec<((NodeId, HashSet<VarAtomBase<A>>), BitAccess)>,
-    arena: &mut SLTNodeArena<A>,
-) -> Result<(NodeId, HashSet<VarAtomBase<A>>), SLTNodeFactsError> {
-    if parts.is_empty() {
-        return Ok((
-            arena.alloc(SLTNode::Constant(
-                BigUint::from(0u32),
-                BigUint::from(0u32),
-                0,
-                false,
-            ))?,
-            HashSet::default(),
-        ));
-    }
-    if parts.len() == 1 {
-        let ((expr, sources), access) = &parts[0];
-        let w = get_width(*expr, arena);
-        if w == 0 {
-            return Ok((*expr, sources.clone()));
-        }
-        if access.lsb == 0 && access.msb == w - 1 {
-            return Ok((*expr, sources.clone()));
-        } else {
-            return Ok((
-                arena.alloc(SLTNode::Slice {
-                    expr: *expr,
-                    access: *access,
-                })?,
-                sources.clone(),
-            ));
-        }
-    }
-
-    let mut concat_parts = Vec::new();
-    let mut total_sources = HashSet::default();
-
-    for ((expr, sources), access) in parts {
-        total_sources.extend(sources);
-        let w = access.msb - access.lsb + 1;
-        let slice = arena.alloc(SLTNode::Slice { expr, access })?;
-        concat_parts.push((slice, w));
-    }
-    concat_parts.reverse();
-    Ok((arena.alloc(SLTNode::Concat(concat_parts))?, total_sources))
-}
-
-#[cfg(test)]
-mod tests {
-    use veryl_analyzer::{
-        Analyzer, Context, attribute_table,
-        ir::{Component, Declaration, Ir, VarPath},
-        symbol_table,
-    };
-    use veryl_metadata::Metadata;
-    use veryl_parser::Parser;
-
-    use super::*;
-    // 既存のインポート...
-    pub struct CombResult {
-        pub paths: Vec<LogicPath<VarId>>,
-        pub boundaries: HashMap<VarId, BTreeSet<usize>>,
-    }
-    pub fn parse_top_module(code: &str) -> Module {
-        symbol_table::clear();
-        attribute_table::clear();
-
-        let metadata = Metadata::create_default("prj").unwrap();
-        let parser = Parser::parse(code, &"").unwrap();
-        let analyzer = Analyzer::new(&metadata);
-        let mut context = Context::default();
-        let mut ir = Ir::default();
-
-        // Pass 1 & 2 を実行して Ir を構築
-        let errors = analyzer.analyze_pass1("prj", &parser.veryl);
-        assert!(errors.is_empty(), "analyze_pass1 errors: {errors:?}");
-        let errors = Analyzer::analyze_post_pass1();
-        assert!(errors.is_empty(), "analyze_post_pass1 errors: {errors:?}");
-        let errors = analyzer.analyze_pass2("prj", &parser.veryl, &mut context, Some(&mut ir));
-        assert!(errors.is_empty(), "analyze_pass2 errors: {errors:?}");
-        let errors = Analyzer::analyze_post_pass2(&ir);
-        assert!(errors.is_empty(), "analyze_post_pass2 errors: {errors:?}");
-
-        // Top モジュールを探す
-        let top_id = veryl_parser::resource_table::insert_str("Top");
-        ir.components
-            .into_iter()
-            .find_map(|e| match e {
-                Component::Module(m) if m.name == top_id => Some(m),
-                _ => None,
-            })
-            .expect("Top module not found")
-    }
-
-    /// 新しい parse_comb の出力を直接検査するためのヘルパー
-    pub fn inspect_comb(code: &str) -> (Module, CombResult) {
-        let top_module = parse_top_module(code);
-
-        // Top モジュール内の最初の always_comb をパース
-        // (実際には複数の場合もあるので、必要に応じて loop させる)
-        let comb_decl = top_module
-            .declarations
-            .iter()
-            .find_map(|d| {
-                if let Declaration::Comb(c) = d {
-                    Some(c)
-                } else {
-                    None
-                }
-            })
-            .expect("No always_comb found in Top");
-        let mut arena = SLTNodeArena::new();
-        let (paths, _, boundaries, _, _) =
-            super::parse_comb(&top_module, comb_decl, &mut arena).unwrap();
-        (top_module, CombResult { paths, boundaries })
-    }
-    pub fn var_id_of(module: &Module, var_path: &[&str]) -> VarId {
-        let mut var_path_str_id = Vec::new();
-        for path in var_path {
-            let id = veryl_parser::resource_table::insert_str(path);
-            var_path_str_id.push(id);
-        }
-        let path = VarPath(var_path_str_id);
-        module
-            .variables
-            .values()
-            .find(|e| e.path == path)
-            .unwrap()
-            .id
-    }
-    #[test]
-    fn test_parse_comb_boundary_collection() {
-        let code = r#"
-            module Top (a: input logic<32>, b: output logic<32>) {
-                               always_comb {
-                    b = 0;
-                    b[7:4] = a[3:0];
-                }
-            }
-        "#;
-        let (module, result) = inspect_comb(code);
-        // 1.① 境界情報が正しく集まっているか
-        let b_id = var_id_of(&module, &["b"]);
-        let bounds = &result.boundaries[&b_id];
-
-        // b[7:4] への代入なので、境界は 4 と 8 が必要
-        assert!(bounds.contains(&4));
-        assert!(bounds.contains(&8));
-
-        // 2. 依存関係の絞り込み (b[7:4] のソースに a[3:0] だけが含まれているか)
-        let path = result
-            .paths
-            .iter()
-            .find(|p| {
-                p.target.var().unwrap().id == b_id
-                    && p.target.var().unwrap().access.lsb == 4
-                    && p.target.var().unwrap().access.msb == 7
-            })
-            .unwrap();
-        let a_id = var_id_of(&module, &["a"]);
-
-        let a_deps: Vec<_> = path.sources.iter().filter(|s| s.id == a_id).collect();
-        assert_eq!(a_deps.len(), 1);
-        assert_eq!(a_deps[0].access.lsb, 0);
-        assert_eq!(a_deps[0].access.msb, 3);
-    }
-
-    #[test]
-    fn test_output_function_body_read_boundaries_propagate() {
-        let code = r#"
-            module Top (a: input logic<8>, q: output logic<4>) {
-                function f (
-                    y: output logic<4>,
-                ) {
-                    y = a[3:0];
-                }
-
-                always_comb {
-                    f(q);
-                }
-            }
-        "#;
-        let (module, result) = inspect_comb(code);
-        let a_id = var_id_of(&module, &["a"]);
-        let bounds = &result.boundaries[&a_id];
-
-        assert!(bounds.contains(&0));
-        assert!(bounds.contains(&4));
-    }
-
-    #[test]
-    fn test_collect_written_accesses_includes_function_call_outputs() {
-        let code = r#"
-            module Top (n: input logic<3>, q: output logic<4>) {
-                function set_bit (
-                    x: input logic,
-                    y: output logic,
-                ) {
-                    y = x;
-                }
-
-                always_comb {
-                    for i in 0..n {
-                        set_bit(1'b0, q[i]);
-                    }
-                }
-            }
-        "#;
-        let module = parse_top_module(code);
-        let comb_decl = module
-            .declarations
-            .iter()
-            .find_map(|d| {
-                if let Declaration::Comb(c) = d {
-                    Some(c)
-                } else {
-                    None
-                }
-            })
-            .expect("No always_comb found in Top");
-        let for_stmt = comb_decl
-            .statements
-            .iter()
-            .find_map(|stmt| {
-                if let Statement::For(for_stmt) = stmt {
-                    Some(for_stmt)
-                } else {
-                    None
-                }
-            })
-            .expect("No for statement found in Top");
-        let mut written = HashMap::default();
-        collect_written_accesses(&module, &for_stmt.body, &mut written).unwrap();
-
-        let q_id = var_id_of(&module, &["q"]);
-        assert_eq!(written[&q_id], vec![BitAccess::new(0, 3)]);
-    }
-
-    #[test]
-    fn test_dependency_override() {
-        let code = r#"
-        module Top (b: input logic<8>, c: input logic<1>, o_a: output logic<8>) {
-            var a: logic<8>;
-            always_comb {
-                a = b;
-                a[0] = c;
-            }
-            assign o_a = a;
-        }
-    "#;
-        let (module, res) = inspect_comb(code);
-        let id_a = var_id_of(&module, &["a"]);
-        let id_b = var_id_of(&module, &["b"]);
-        let id_c = var_id_of(&module, &["c"]);
-
-        // Find path for a[0]
-        let path_a0 = res
-            .paths
-            .iter()
-            .find(|p| p.target.var().unwrap().id == id_a && p.target.var().unwrap().access.lsb == 0)
-            .expect("Path for a[0] not found");
-
-        // a[0] depends on c
-        assert!(
-            path_a0.sources.iter().any(|s| s.id == id_c),
-            "a[0] must depend on c"
-        );
-        // a[0] should NOT depend on b
-        assert!(
-            !path_a0.sources.iter().any(|s| s.id == id_b),
-            "a[0] must NOT depend on b"
-        );
-
-        // Find path for a[7:1]
-        let path_a_upper = res
-            .paths
-            .iter()
-            .find(|p| p.target.var().unwrap().id == id_a && p.target.var().unwrap().access.lsb == 1)
-            .expect("Path for a[7:1] not found");
-        assert!(
-            path_a_upper.sources.iter().any(|s| s.id == id_b),
-            "a[7:1] must depend on b"
-        );
-    }
-
-    #[test]
-    fn test_arithmetic_dependency() {
-        let code = r#"
-        module Top (b: input logic<8>, c: input logic<8>, o_a: output logic<8>) {
-            assign o_a = b + c;
-        }
-    "#;
-        let (module, res) = inspect_comb(code);
-        let id_oa = var_id_of(&module, &["o_a"]);
-        let id_b = var_id_of(&module, &["b"]);
-        let id_c = var_id_of(&module, &["c"]);
-
-        let path_oa = res
-            .paths
-            .iter()
-            .find(|p| p.target.var().unwrap().id == id_oa)
-            .unwrap();
-
-        // o_a depends on b and c
-        assert!(path_oa.sources.iter().any(|s| s.id == id_b));
-        assert!(path_oa.sources.iter().any(|s| s.id == id_c));
-    }
-
-    #[test]
-    fn test_div_rem_parser_selects_explicit_signed_variants() {
-        let code = r#"
-            module Top (
-                ua: input logic<8>,
-                ub: input logic<8>,
-                sa: input signed logic<8>,
-                sb: input signed logic<8>,
-                udiv: output logic<8>,
-                urem: output logic<8>,
-                sdiv: output signed logic<8>,
-                srem: output signed logic<8>,
-                mixed: output logic<8>
-            ) {
-                always_comb {
-                    udiv = ua / ub;
-                    urem = ua % ub;
-                    sdiv = sa / sb;
-                    srem = sa % sb;
-                    mixed = sa / ub;
-                }
-            }
-        "#;
-        let module = parse_top_module(code);
-        let comb_decl = module
-            .declarations
-            .iter()
-            .find_map(|declaration| match declaration {
-                Declaration::Comb(declaration) => Some(declaration),
-                _ => None,
-            })
-            .unwrap();
-        let mut arena = SLTNodeArena::new();
-        let (paths, _, _, _, _) = super::parse_comb(&module, comb_decl, &mut arena).unwrap();
-
-        let op_for = |name| {
-            let target = var_id_of(&module, &[name]);
-            let path = paths
-                .iter()
-                .find(|path| path.target.var().is_some_and(|var| var.id == target))
-                .unwrap();
-            match arena.get(path.expr) {
-                SLTNode::Binary(_, op, _) => *op,
-                node => panic!("expected binary root for {name}, got {node:?}"),
-            }
-        };
-
-        assert_eq!(op_for("udiv"), BinaryOp::DivU);
-        assert_eq!(op_for("urem"), BinaryOp::RemU);
-        assert_eq!(op_for("sdiv"), BinaryOp::DivS);
-        assert_eq!(op_for("srem"), BinaryOp::RemS);
-        assert_eq!(op_for("mixed"), BinaryOp::DivU);
-    }
-
-    #[test]
-    fn test_bit_level_self_assignment_dag() {
-        let code = r#"
-        module Top (i: input logic<8>, o: output logic<8>) {
-            var a: logic<8>;
-            always_comb {
-                a = i;
-                a[0] = a[1];
-            }
-            assign o = a;
-        }
-    "#;
-        let (module, res) = inspect_comb(code);
-        let id_a = var_id_of(&module, &["a"]);
-        let id_i = var_id_of(&module, &["i"]);
-
-        // a[0] = a[1] = i[1]
-        let path_a0 = res
-            .paths
-            .iter()
-            .find(|p| p.target.var().unwrap().id == id_a && p.target.var().unwrap().access.lsb == 0)
-            .unwrap();
-
-        assert!(
-            path_a0
-                .sources
-                .iter()
-                .any(|s| s.id == id_i && s.access.lsb <= 1 && s.access.msb >= 1),
-            "a[0] should depend on i[1]"
-        );
-    }
-    #[test]
-    fn test_dynamic_assign_eval() {
-        let code = r#"
-            module Top (
-                a: input logic<32>,
-                idx: input logic<5>,
-                val: input logic<1>,
-                d: output logic<32>
-            ) {
-                always_comb {
-                    d = a;
-                    d[idx] = val;
-                }
-            }
-        "#;
-        let (module, result) = inspect_comb(code);
-
-        // d is updated dynamically, so we expect a path covering d[0..31]
-        let id_d = var_id_of(&module, &["d"]);
-        let path = result
-            .paths
-            .iter()
-            .find(|p| p.target.var().unwrap().id == id_d);
-
-        // Dynamic assignment essentially combines all bits, so we should find a path for d.
-        // It might be split or single, but since we updated full range in eval_dynamic_assign, it should be single if initialized so.
-        // But `d=a` initializes it with 0..31 (or splits if `a` is split). `a` is input 32.
-        // So `d` starts as [0:31]. Dynamic update updates [0:31]. So it should stay [0:31].
-
-        let path = path.expect("Path for d not found");
-        assert_eq!(path.target.var().unwrap().access.lsb, 0);
-        assert_eq!(path.target.var().unwrap().access.msb, 31);
-
-        let id_a = var_id_of(&module, &["a"]);
-        let id_idx = var_id_of(&module, &["idx"]);
-        let id_val = var_id_of(&module, &["val"]);
-
-        assert!(
-            path.sources.iter().any(|s| s.id == id_a),
-            "Depends on old value a"
-        );
-        assert!(
-            path.sources.iter().any(|s| s.id == id_idx),
-            "Depends on index idx"
-        );
-        assert!(
-            path.sources.iter().any(|s| s.id == id_val),
-            "Depends on new value val"
-        );
-    }
-
-    #[test]
-    fn test_slt_display() {
-        let mut arena = SLTNodeArena::<i32>::new();
-        // Test simple constant
-        let _const_node = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(42u32),
-                BigUint::from(0u32),
-                8,
-                false,
-            ))
-            .unwrap();
-        // fmt_display is not easily callable here without a Formatter, but we can check if it compiles or use a dummy formatter
-        // Actually, let's just use a custom wrapper with Display if needed, but for now let's just fix the test to compile.
-
-        // Test unary operation
-        let inner = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(5u32),
-                BigUint::from(0u32),
-                4,
-                false,
-            ))
-            .unwrap();
-        let _unary_node = arena.alloc(SLTNode::Unary(UnaryOp::Minus, inner)).unwrap();
-
-        // Test binary operation
-        let lhs = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(1u32),
-                BigUint::from(0u32),
-                8,
-                false,
-            ))
-            .unwrap();
-        let rhs = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(2u32),
-                BigUint::from(0u32),
-                8,
-                false,
-            ))
-            .unwrap();
-        let _binary_node = arena
-            .alloc(SLTNode::Binary(lhs, BinaryOp::Add, rhs))
-            .unwrap();
-
-        // Test Mux
-        let cond = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(1u32),
-                BigUint::from(0u32),
-                1,
-                false,
-            ))
-            .unwrap();
-        let then_expr = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(10u32),
-                BigUint::from(0u32),
-                8,
-                false,
-            ))
-            .unwrap();
-        let else_expr = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(20u32),
-                BigUint::from(0u32),
-                8,
-                false,
-            ))
-            .unwrap();
-        let _mux_node = arena
-            .alloc(SLTNode::Mux {
-                cond,
-                then_expr,
-                else_expr,
-            })
-            .unwrap();
-
-        // Test Concat
-        let parts = vec![
-            (
-                arena
-                    .alloc(SLTNode::Constant(
-                        BigUint::from(1u32),
-                        BigUint::from(0u32),
-                        4,
-                        false,
-                    ))
-                    .unwrap(),
-                4,
-            ),
-            (
-                arena
-                    .alloc(SLTNode::Constant(
-                        BigUint::from(2u32),
-                        BigUint::from(0u32),
-                        4,
-                        false,
-                    ))
-                    .unwrap(),
-                4,
-            ),
-        ];
-        let _concat_node = arena.alloc(SLTNode::Concat(parts)).unwrap();
-
-        // Test Slice
-        let expr = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(255u32),
-                BigUint::from(0u32),
-                8,
-                false,
-            ))
-            .unwrap();
-        let _slice_node = arena
-            .alloc(SLTNode::Slice {
-                expr,
-                access: BitAccess::new(2, 5),
-            })
-            .unwrap();
-    }
-
-    #[test]
-    fn test_slt_display_complex() {
-        let mut arena = SLTNodeArena::<i32>::new();
-        // Display complex nested expression: (a + b) * (c - d)
-        let a = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(1u32),
-                BigUint::from(0u32),
-                32,
-                false,
-            ))
-            .unwrap();
-        let b = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(2u32),
-                BigUint::from(0u32),
-                32,
-                false,
-            ))
-            .unwrap();
-        let add_expr = arena.alloc(SLTNode::Binary(a, BinaryOp::Add, b)).unwrap();
-
-        let c = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(3u32),
-                BigUint::from(0u32),
-                32,
-                false,
-            ))
-            .unwrap();
-        let d = arena
-            .alloc(SLTNode::Constant(
-                BigUint::from(4u32),
-                BigUint::from(0u32),
-                32,
-                false,
-            ))
-            .unwrap();
-        let sub_expr = arena.alloc(SLTNode::Binary(c, BinaryOp::Sub, d)).unwrap();
-
-        let _mul_node = arena
-            .alloc(SLTNode::Binary(add_expr, BinaryOp::Mul, sub_expr))
-            .unwrap();
-    }
-
-    #[test]
-    fn loop_bound_status_allows_exclusive_upper_sentinel() {
-        assert_eq!(
-            super::loop_bound_status(&ForBound::Const(255), 8, false),
-            Some(super::LoopBoundStatus::FitsLoopType)
-        );
-        assert_eq!(
-            super::loop_bound_status(&ForBound::Const(256), 8, false),
-            Some(super::LoopBoundStatus::ExclusiveUpperSentinel)
-        );
-        assert_eq!(
-            super::loop_bound_status(&ForBound::Const(257), 8, false),
-            Some(super::LoopBoundStatus::OutOfRange)
-        );
-    }
-}
+pub use celox_frontend_veryl::logic_tree::*;
