+++ docs/internals/compiler-crate-architecture.md
@@ -39,8 +39,9 @@ symbolic, optimizer, and runtime payload into the phase-specific target crates b
 `celox-slt` now owns the frontend-independent bit-range store, symbolic-state contract, SLT node
 arena, iterative facts/verifier, path graph, fused comb/FF scheduler, and SLT-to-SIR lowerer. The FF
 lowering callback has an associated frontend-owned error type, so the scheduler no longer imports
-parser errors. Veryl AST traversal remains in the facade as an arena builder and is the remaining
-SLT construction move.
+parser errors. `celox-frontend-veryl` owns lowering diagnostics, context-width and bit-select
+semantics, analyzer loop provenance, and the Veryl comb arena builder. Module/FF orchestration is
+the remaining frontend construction move.
 
 The baseline is the compiler pipeline on `perf/native-simulation-throughput` after PR #322. The
 split must preserve RTL semantics, generated-code quality, and the public `celox` API while making
